### 2023-09-11 #5803 @eloisabentivegna
@Sbozzolo, @rhaas80: this submission appears to be very close to your interests, and your feedback would be much appreciated. Would you be available to review?
