+++ README.md
@@ -1,3 +1,5 @@
+[![DOI](https://zenodo.org/badge/685792148.svg)](https://zenodo.org/badge/latestdoi/685792148)
+
 # mayawaves
 
 `Mayawaves` serves as an analysis library for Einstein Toolkit and MAYA simulations. Follow the tutorials and documentation at https://mayawaves.github.io/mayawaves to learn how
