### 2023-09-14 #5803 @eloisabentivegna
> possibly, but I would be somewhat conflicted I think since Deborah just started as a postdoc in Helvi's group at NCSA (and is co-funded by the ET grant at NCSA) and I was involved in the hiring process. So I can comment,but not much else.

@rhaas80, thanks for disclosing the conflict. I agree that it would be best for you not to be involved with the review itself. I am still interested in your comments, though. I will wait until the reviewers express their recommendations and then mention you in the review issue to solicit your input. Is this OK?
