### 2023-09-20 bb6957c
Add automated tests (#6)
