+++ .github/ci.yml
@@ -1,7 +1,7 @@
 name: Test Suite
 on:
   pull_request:
-    branches: [ master ]
+    branches: [ main ]
 
 jobs:
   build:
