+++ .github/workflows/test_suite.yml
@@ -9,8 +9,8 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
-      	with:
-	  python-version: '3.10' 
+        with:
+          python-version: '3.10' 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
