### 2023-09-20 44edfdf
modify 2 tests to be isclose rather than equal since numerical precision caused some tests to fail
