### 2023-09-20 1609bad
Update test of attributes
