### 2023-09-21 b520425
Update tests and inspiral crop method
