+++ mayawaves/utils/postprocessingutils.py
@@ -3047,15 +3047,21 @@ def summarize_coalescence(coalescence: Coalescence, output_directory: str = None
 
     # Format information to be printed to summary.txt:
     # beginning includes information regarding initial data.
+    mass_ratio_string = 'NaN' if initial_mass_ratio is None else f'{initial_mass_ratio:.4f}'
+    primary_horizon_mass_string = 'NaN' if initial_primary_horizon_mass is None else f'{initial_primary_horizon_mass:.4f}'
+    secondary_horizon_mass_string = 'NaN' if initial_secondary_horizon_mass is None else f'{initial_secondary_horizon_mass:.4f}'
+    primary_spin_string = 'NaN' if initial_primary_dimensionless_spin is None else f'[{initial_primary_dimensionless_spin[0]:.4f}, {initial_primary_dimensionless_spin[1]:.4f}, {initial_primary_dimensionless_spin[2]:.4f}]'
+    secondary_spin_string = 'NaN' if initial_secondary_dimensionless_spin is None else f'[{initial_secondary_dimensionless_spin[0]:.4f}, {initial_secondary_dimensionless_spin[1]:.4f}, {initial_secondary_dimensionless_spin[2]:.4f}]'
+    separation_string = 'NaN' if initial_separation is None else f'{initial_separation:.4f}'
     beginning = f"Summary for {coalescence_name}\n" \
                 f"\nInitial Values (t=0 M):\n" \
                 f"{'-' * 70}\n" \
-                f"mass ratio:\t\t\t\t{str(initial_mass_ratio)[0:6]}\n" \
-                f"primary horizon mass:\t\t\t{str(initial_primary_horizon_mass)[0:6]}\n" \
-                f"secondary horizon mass:\t\t\t{str(initial_secondary_horizon_mass)[0:6]}\n" \
-                f"primary dimensionless spin:\t\t{[round(i, 4) for i in initial_primary_dimensionless_spin]}\n" \
-                f"secondary dimensionless spin:\t\t{[round(i, 4) for i in initial_secondary_dimensionless_spin]}\n" \
-                f"separation:\t\t\t\t{str(initial_separation)[0:6]} M\n"
+                f"mass ratio:\t\t\t\t{mass_ratio_string}\n" \
+                f"primary horizon mass:\t\t\t{primary_horizon_mass_string}\n" \
+                f"secondary horizon mass:\t\t\t{secondary_horizon_mass_string}\n" \
+                f"primary dimensionless spin:\t\t{primary_spin_string}\n" \
+                f"secondary dimensionless spin:\t\t{secondary_spin_string}\n" \
+                f"separation:\t\t\t\t{separation_string} M\n"
 
     # middle includes information regarding data at t=75 M.
     if current_time < 75:
@@ -3068,40 +3074,53 @@ def summarize_coalescence(coalescence: Coalescence, output_directory: str = None
                      f"{'-' * 70}\n" \
                      f"Simulation completed before reaching t=75 M.\n"
     else:
-        pds_message = f"None" if primary_dimensionless_spin is None else f"{[round(i, 4) for i in primary_dimensionless_spin]}"
-        sds_message = f"None" if secondary_dimensionless_spin is None else f"{[round(i, 4) for i in secondary_dimensionless_spin]}"
+        primary_horizon_mass_string = 'NaN' if primary_horizon_mass is None else f'{primary_horizon_mass:.4f}'
+        secondary_horizon_mass_string = 'NaN' if secondary_horizon_mass is None else f'{secondary_horizon_mass:.4f}'
[…]
