+++ tests/test_postprocessingutils.py
@@ -3468,29 +3468,29 @@ class TestPostprocessingUtils(TestCase):
 
 Initial Values (t=0 M):
 ----------------------------------------------------------------------
-mass ratio:				1.9999
+mass ratio:				2.0000
 primary horizon mass:			0.6665
-secondary horizon mass:			0.3332
-primary dimensionless spin:		[0.0, 0.0, 0.4001]
-secondary dimensionless spin:		[0.0, 0.0, 0.4]
-separation:				11.000 M
+secondary horizon mass:			0.3333
+primary dimensionless spin:		[0.0000, 0.0000, 0.4001]
+secondary dimensionless spin:		[0.0000, 0.0000, 0.4000]
+separation:				11.0001 M
 
 After Junk Radiation (t=75 M):
 ----------------------------------------------------------------------
 primary horizon mass:			0.6667
 secondary horizon mass:			0.3334
-primary dimensionless spin:		[-0.0, 0.0, 0.4007]
-secondary dimensionless spin:		[-0.0, 0.0, 0.402]
-separation:				10.778 M
-separation unit vector:			[0.1005, -0.9949, -0.0]
+primary dimensionless spin:		[-0.0000, 0.0000, 0.4007]
+secondary dimensionless spin:		[-0.0000, 0.0000, 0.4020]
+separation:				10.7785 M
+separation unit vector:			[0.1005, -0.9949, -0.0000]
 orbital frequency:			0.0243
-orbital angular momentum unit vector:	[0.0, 0.0, 1.0]
+orbital angular momentum unit vector:	[0.0000, 0.0000, 1.0000]
 eccentricity:				0.0034
 
 Remnant Values (Last available data):
 ----------------------------------------------------------------------
 horizon mass:				0.9492
-dimensionless spin:			[-0.0, 0.0, 0.7701]"""
+dimensionless spin:			[-0.0000, 0.0000, 0.7701]"""
 
         summary_file = open(path_to_summary_file, "r")
         actual_output = summary_file.read()
@@ -3540,30 +3540,30 @@ dimensionless spin:			[-0.0, 0.0, 0.7701]"""
 
 Initial Values (t=0 M):
 ----------------------------------------------------------------------
-mass ratio:				1.0
-primary horizon mass:			0.5167
-secondary horizon mass:			0.5167
-primary dimensionless spin:		[0.0, 0.0, 0.0006]
-secondary dimensionless spin:		[0.0, 0.0, 0.0006]
+mass ratio:				1.0000
+primary horizon mass:			0.5168
+secondary horizon mass:			0.5168
+primary dimensionless spin:		[0.0000, 0.0000, 0.0006]
+secondary dimensionless spin:		[0.0000, 0.0000, 0.0006]
 separation:				2.3363 M
 
 After Junk Radiation (t=75 M):
 ----------------------------------------------------------------------
-primary horizon mass:			None
-secondary horizon mass:			None
-primary dimensionless spin:		None
-secondary dimensionless spin:		None
-separation:				1.0170 M
-separation unit vector:			[-0.1809, -0.9835, 0.0]
+primary horizon mass:			NaN
+secondary horizon mass:			NaN
+primary dimensionless spin:		NaN
+secondary dimensionless spin:		NaN
+separation:				0.0000 M
+separation unit vector:			[-0.1809, -0.9835, 0.0000]
 orbital frequency:			0.4345
-orbital angular momentum unit vector:	[0.0, 0.0, 1.0]
+orbital angular momentum unit vector:	[0.0000, 0.0000, 1.0000]
 eccentricity:				0.0243
-**A 'None' value represents data that was not being tracked at t=75 M.**
+**A 'NaN' value represents data that was not being tracked at t=75 M.**
 
 Remnant Values (Last available data):
[…]
