### 2023-09-21 2baa45a
Fix yml (#10)
* Put yml in the right folder

* Change checkout version

* Add python version

* change tabs to spaces

* modify 2 tests to be isclose rather than equal since numerical precision caused some tests to fail

* Update test of attributes

* Update tests and inspiral crop method

---------
