+++ docs/html/_sources/source/mayawaves.radiation.rst.txt
@@ -1,22 +1,11 @@
 mayawaves.radiation module
 ==========================
 
-mayawaves.radiation.RadiationBundle class
------------------------------------------
-
 .. autoclass:: mayawaves.radiation.RadiationBundle
    :members:
 
-
-mayawaves.radiation.RadiationSphere class
------------------------------------------
-
 .. autoclass:: mayawaves.radiation.RadiationSphere
    :members:
-
-
-mayawaves.radiation.RadiationMode class
------------------------------------------
 
 .. autoclass:: mayawaves.radiation.RadiationMode
    :members:
