### 2023-10-23 17021ae
reduce filesizes of sample simulation
