+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "mayawaves"
-version = "2023.8"
+version = "2023.10"
 authors = [
     { name="Deborah Ferguson", email="deborah.l.ferguson@gmail.com" },
     { name="Surendra Anne" },
