+++ example_simulation/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/output-0002/SIMFACTORY/properties.ini
@@ -1,34 +1,0 @@
-
-[properties]
-machine         = stampede2-skx
-simulationid    = simulation-D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33-stampede2-skx-stampede2.tacc.utexas.edu-ryan_n16-2022.09.30-14.39.03-21772
-sourcedir       = /work2/06913/ryan_n16/stampede2/Cactus
-configuration   = bbh-maya-automated
-configid        = config-bbh-maya-automated-login4.stampede2.tacc.utexas.edu-work2-06913-ryan_n16-stampede2-Cactus
-buildid         = build-bbh-maya-automated-login4.stampede2.tacc.utexas.edu-ryan_n16-2022.09.16-20.52.45-252857
-testsuite       = False
-executable      = /scratch/06913/ryan_n16/simulations/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/SIMFACTORY/exe/cactus_bbh-maya-automated
-optionlist      = /scratch/06913/ryan_n16/simulations/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/SIMFACTORY/cfg/OptionList
-submitscript    = /scratch/06913/ryan_n16/simulations/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/SIMFACTORY/run/SubmitScript
-runscript       = /scratch/06913/ryan_n16/simulations/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/SIMFACTORY/run/RunScript
-parfile         = /scratch/06913/ryan_n16/simulations/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33/SIMFACTORY/par/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33.rpar
-chainedjobid    = 10367603
-ppn             = 48
-procsrequested  = 1056
-allocation      = TG-PHY120016
-user            = ryan_n16
-numsmt          = 1
-walltime        = 24:00:00
-numprocs        = 264
-nodeprocs       = 12
-numthreads      = 2
-hostname        = stampede2.tacc.utexas.edu
-ppnused         = 24
-queue           = skx-normal
-cpufreq         = 2.1
-procs           = 528
-memory          = 196608
-nodes           = 22
-pbsSimulationName= D11_q5_a1_-0.36
-jobid           = 10379567
-checkpointing   = yes
