+++ tests/resources/main_test_simulation/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/output-0001-active/SIMFACTORY/properties.ini
@@ -1,34 +1,0 @@
-
-[properties]
-machine         = frontera-clx
-simulationid    = simulation-D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67-frontera-clx-frontera.tacc.utexas.edu-dferg-2021.01.21-22.28.29-236870
-sourcedir       = /work/05765/dferg/frontera/Cactus
-configuration   = bbh-maya
-configid        = config-bbh-maya-login1.frontera.tacc.utexas.edu-work-05765-dferg-frontera-Cactus
-buildid         = build-bbh-maya-login1.frontera.tacc.utexas.edu-dferg-2020.11.18-22.06.18-145326
-testsuite       = False
-executable      = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/exe/cactus_bbh-maya
-optionlist      = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/cfg/OptionList
-submitscript    = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/run/SubmitScript
-runscript       = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/run/RunScript
-parfile         = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/par/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67.rpar
-chainedjobid    = -1
-ppn             = 56
-procsrequested  = 56
-allocation      = PHY20039
-user            = dferg
-numsmt          = 1
-walltime        = 4:00:00
-numprocs        = 56
-nodeprocs       = 56
-numthreads      = 1
-hostname        = frontera.tacc.utexas.edu
-ppnused         = 56
-queue           = normal
-cpufreq         = 2.7
-procs           = 56
-memory          = 196608
-nodes           = 1
-pbsSimulationName= D2.33_q1_a1_0_0
-jobid           = 2402004
-checkpointing   = yes
