### 2023-10-23 ccd115a
New build (#12)
* reduce filesizes of sample simulation

* reduce test files

* reduce file sizes

---------
