+++ CITATION.cff
@@ -30,7 +30,7 @@ authors:
   - family-names: "Valorz"
     given-names: "Finny"
 title: "mayawaves"
-version: v2023.8
-date-released: 2023-09-12
+version: v2023.10
+date-released: 2023-10-23
 url: "https://github.com/MayaWaves/mayawaves"
 doi: 10.5281/zenodo.8339253

[…]
