### 2023-10-23 69072ae
Update citation.cff (#14)
Co-authored-by: Deborah Ferguson <deborahferguson@wirelessprv-10-194-86-5.near.illinois.edu>
