+++ CITATION.cff
@@ -33,4 +33,4 @@ title: "mayawaves"
 version: v2023.10
 date-released: 2023-10-23
 url: "https://github.com/MayaWaves/mayawaves"
-doi: 10.5281/zenodo.8339253
+doi: 10.5281/zenodo.10035526
