### 2023-10-28 #5803 @barrywardell
> @barrywardell, you have also worked on analysis packages for the ET. Would you be available to review this submission?

In principle I would be happy to do the review, but unfortunately I'm right in the middle of a busy teaching semester so I don't foresee having any time for at least another ~6 weeks. If you're still looking for a reviewer at that point please let me know.
