### 2023-11-07 #6032 @editorialbot
[REVIEW]: Mayawaves: Python Library for Interacting with the Einstein Toolkit and the MAYA Catalog
**Submitting author:** @deborahferguson (<a href="http://orcid.org/0000-0002-4406-591X">Deborah Ferguson</a>)
**Repository:** https://github.com/MayaWaves/mayawaves
**Branch with paper.md** (empty if default branch): paper
**Version:** v2024.6
