### 2023-11-08 a5d12be
Add missing DOI/fix reference
