### 2023-11-08 9a6f6e3
Add missing DOI/fix reference
