+++ paper.bib
@@ -310,14 +310,18 @@
     year = "2017"
 }
 
-@inproceedings{Herrmann:2006ks,
-    author = "Herrmann, F. and Hinder, Ian and Shoemaker, D. and Laguna, P.",
-    title = "{Unequal Mass Binary Black Hole Plunges and Gravitational Recoil}",
-    booktitle = "{New Frontiers in Numerical Relativity (NFNR 2006)}",
-    eprint = "gr-qc/0601026",
-    archivePrefix = "arXiv",
-    month = "1",
-    year = "2006"
+@article{Herrmann_2007,
+    doi = {10.1088/0264-9381/24/12/S04},
+    url = {https://dx.doi.org/10.1088/0264-9381/24/12/S04},
+    year = {2007},
+    month = {may},
+    publisher = {},
+    volume = {24},
+    number = {12},
+    pages = {S33},
+    author = {Frank Herrmann and Ian Hinder and Deirdre Shoemaker and Pablo Laguna},
+    title = {Unequal mass binary black hole plunges and gravitational recoil},
+    journal = {Classical and Quantum Gravity}
 }
 
 @article{Jani:2016wkt,
