### 2023-11-11 #6032 @Sbozzolo
Hi @deborahferguson, given the long time it took to find you an editor and reviewers, I already started looking at the repository and the paper. There are some important issues that would be good to address first (before the rest of the review can proceed).

My first impression is that the documentation is rather minimal (as far I can see there's only a few jupyter notebooks and the function docstrings).

Importantly:

- I couldn't find how to install the package, or what versions of Python are supported. I looked at the pyproject.toml and didn't recognize the build system, so I couldn't get started with the package.
- It is not clear what versions of Einstein Toolkit/thorns are supported.
- There's no information about how to report problems, how to contribute to the package, how to ask for help, or how to run tests.

In addition to this, it would be helpful to expand the documentation to discuss the various modules in more details, provide references/equations/conventions used, discuss features, et cetera.

Other first impressions:

The git repository is quite large (560 MB), entirely due to the files in the test folder (the code is 500 kB). It would be good to see if it is possible to reduce the size.

The git history is almost empty, so I cannot judge about authorship of the software.

Regarding the paper, it is well written. Here's some recommendations for content that should be included:

I think it would be good to mention that Einstein Toolkit is open source (is maya open source)? Given the fragmented nature of Einstein Toolkit, it would be useful to explicitly mention which thorns are supported. The paper also does not survey the state of the art regarding packages for gravitational wave analysis in Einstein Toolkit. In particular, it would be important to to discuss what additional features mayawaves has compared to the two gravitational wave analysis Python packages that come with Einstein Toolkit: kuibit (of which I am author), and POWER.

I am looking forward to reviewing your package!
