### 2023-11-14 #6032 @deborahferguson
@Sbozzolo In regards to the authorship, this was developed on an internal GitHub and when we decided to make it public, I moved it onto the normal GitHub. I chose not to port over all the git history due to security risks of early versions having local file paths, etc hardcoded into files (particularly in regards to tests and example simulations). That git history does still exist so maybe I can share screenshots or something of the contribution statistics
