### 2023-11-18 #6032 @eloisabentivegna
Good point, @Sbozzolo. Notice that non-code contributions can also be grounds for authorship in JOSS (see https://joss.readthedocs.io/en/latest/submitting.html#authorship). So, if the author list cannot clearly be established by inspecting the code repo, it is ultimately up to @deborahferguson to state who should be included (and perhaps provide a quick explanation why).

Having said this, I would like to loop in @openjournals/aass-eics to confirm.
