### 2023-11-20 #6032 @dfm
> Having said this, I would like to loop in @openjournals/aass-eics to confirm.

Thanks @eloisabentivegna! I agree with everything you've said here about our authorship policy. I think that given the unique context here (missing git history) it is useful to call this out, but we're generally happy to proceed given a summary like @deborahferguson's above. Thanks all!
