### 2023-11-22 #6032 @cjoana
Hi @deborahferguson , I am still testing these packages but so far they look great.  A very useful package to analyse binary black hole mergers simulations and extract data from them. I understand that the current packages is dedicated to only black holes, but not for simulations with  other type of compact objects such as neutron stars.

*  During installation: First I tried to pip install  the provided requirements.txt. I was using python 3.12 and failed because of dependencies. Then, after looking at the error,  I tried  python 3.8 and work well.  You might want to test and document this in the documentation.

 I did some tests on the provided python notebooks. I understand you are currently working on expanding the documentation so I will not comment much on this, but only add some minor comments:
* All notebooks uses the h5 file created from `creating_h5.ipynb`. You might want to add a note on that (in the notebooks), because this file doesn’t exist initially, so the user should first run that notebook to be able to use this tutorial.
* As mention by SBozzolo, the simulation example is rather heavy. So a suggestion would be to store this data somewhere else and provide some inline code to download it.  “Somewhere else” could be a different GitHub repository, Zenodo, external website, etc.  Adding in the documentation some basic description and the code source of the provided example would also be appreciated.
* In `exporting_lvk.ipynb` :  There is a typo in the variable name `output_direcory`  (missing a t in directory).
* In `gravitational_waves.ipynb` :  For consistency with the other notebooks, you might want to replace the name variable `example_h5_filename` by `example_h5_filepath`.
* In the `Readme.txt` of the repository, there is a typo in 4th paragraph in “tuttorial.” ( with two ’t’).

Overall, I found the notebooks useful and well structured to understand and a useful tool to guide on how to use these packages. I will continue testing them with alternative data from the Maya website.  Regarding the paper, I also found it well written and easy to follow.

I am looking forward to the seeing the new additions on the documentation.
