### 2023-12-02 #6032 @Sbozzolo
@deborahferguson

I learned from the comment above that I can install mayawaves with pip, so I did so. (I had to change one constraint to use python 3.11).

Now, I am trying to follow the first tutorial with some data generated with the Einstein Toolkit, but I cannot get past the first step. I've tried with a few different datasets I have, including the clean GW150914 example, and got a few different errors. In some cases, I can understand that the errors come from the fact that the data was generated with modified versions of the Einstein Toolkit and I saw in the mayawaves assume a very specific structure for files/column numbering.
In other cases, I can understand that the errors come from the fact that the dataset I have on my computer is not complete or is in in a non-SIMFACTORY folder structure. For example, some of my datasets do not have the SIMFACTORY folder and par files cannot be found (even if par files exist somewhere else in the simulation output).

The GW150914 case comes directly from Einstein Toolkit, and the error is:
```python
In [2]: mkdir /tmp/tmp3

In [3]: output = "/tmp/tmp3"

In [4]: from mayawaves.utils.postprocessingutils import create_h5_from_simulation

In [5]: simulation = "../GW150914/"

In [6]: h5_path = create_h5_from_simulation(simulation, output)
storing parameter file
storing radiative information
storing compact object information
storing metadata
KeyError                                  Traceback (most recent call last)
Cell In[6], line 1
----> 1 h5_path = create_h5_from_simulation(simulation, output)

File ~/Downloads/mayawaves/mayawaves/utils/postprocessingutils.py:2004, in create_h5_from_simulation(raw_directory, output_directory, catalog_id)
   2002 # process metadata
   2003 print("storing metadata")
-> 2004 _store_meta_data(h5_file, relevant_data_filepaths, relevant_output_filepaths,
   2005                  simulation_name, catalog_id)
   2007 # close h5file
   2008 h5_file.close()

File ~/Downloads/mayawaves/mayawaves/utils/postprocessingutils.py:1926, in _store_meta_data(h5_file, relevant_data_filepaths, relevant_output_filepaths, simulation_name, catalog_id)
   1924     h5_file.attrs["catalog id"] = catalog_id
   1925 # initial dimensional spins
-> 1926 dimensional_spin_0, dimensional_spin_1 = _get_initial_dimensional_spins(h5_file)
   1927 h5_file.attrs["dimensional spin 0"] = dimensional_spin_0
   1928 h5_file.attrs["dimensional spin 1"] = dimensional_spin_1

File ~/Downloads/mayawaves/mayawaves/utils/postprocessingutils.py:1541, in _get_initial_dimensional_spins(h5_file)
   1528 def _get_initial_dimensional_spins(h5_file: h5py.File) -> tuple:
   1529     """Initial dimensional spins.
   1530
   1531     Tries to get the initial dimensional spins from the compact object data. Otherwise, returns the spins set in the
   (...)
   1539
   1540     """
-> 1541     compact_object_data_0 = h5_file["compact_object"]["object=0"]
   1542     compact_object_data_1 = h5_file["compact_object"]["object=1"]
   1544     try:

File h5py/_objects.pyx:54, in h5py._objects.with_phil.wrapper()

File h5py/_objects.pyx:55, in h5py._objects.with_phil.wrapper()

File ~/.cache/pypoetry/virtualenvs/mayawaves-5r_dnVHS-py3.11/lib/python3.11/site-packages/h5py/_hl/group.py:357, in Group.__getitem__(self, name)
    355         raise ValueError("Invalid HDF5 object reference")
    356 elif isinstance(name, (bytes, str)):
--> 357     oid = h5o.open(self.id, self._e(name), lapl=self._lapl)
    358 else:
    359     raise TypeError("Accessing a group is done with bytes or str, "
    360                     "not {}".format(type(name)))

File h5py/_objects.pyx:54, in h5py._objects.with_phil.wrapper()

File h5py/_objects.pyx:55, in h5py._objects.with_phil.wrapper()

File h5py/h5o.pyx:189, in h5py.h5o.open()

KeyError: "Unable to synchronously open object (object 'object=0' doesn't exist)"

```

As you improve the documentation, it would be very useful to clearly collect and present all the requirements to use the code and the assumptions made (also for the physics).

Finally, it would be useful to wrap your HDF5 operations in a context manager to ensure that the file is properly closed when errors occurs. I had to keep restarting the repl or changing output directory everytime an error occurred because of `OSError: Unable to synchronously create file (unable to truncate a file which is already open)`
