### 2023-12-04 #6032 @deborahferguson
Hi @Sbozzolo ,

Thanks for trying to get this working! I just ran it on a clean run of the GW150914 example (just one output, so not a complete run) using the pip version of mayawaves and it worked for me. I am submitting with simfactory and I now realize I've never tested to see if this works without it as our group has always used simfactory for submission. I'll take a look at what modifications need to happen for it to work without simfactory. Thanks for catching that! And yes, I'll add more documentation relating to these things. Apologies that I'm behind on that; it's application/grant season, but I've finally come back to life a bit this week.
