### 2023-12-04 #6032 @deborahferguson
Hi @cjoana

I'm glad you're liking it so far! Yes, I am working on improving the documentation and the python version is definitely important. Unfortunately one of the libraries I'm using to perform frame conversions isn't compatible beyond python 3.10 if I recall. I'll definitely add that to the documentation. I like your solution for reducing the size! I didn't like that it was so heavy but I also wanted to be able to include a full sample simulation. Hosting it somewhere else is a great idea. Thanks for catching the typos too!
