### 2023-12-04 #6032 @Sbozzolo
> Hi @Sbozzolo ,
>
> Thanks for trying to get this working! I just ran it on a clean run of the GW150914 example (just one output, so not a complete run) using the pip version of mayawaves and it worked for me.

Have you tried with the GW150914 data from [Zenodo](https://zenodo.org/records/155394)?

> I am submitting with simfactory and I now realize I've never tested to see if this works without it as our group has always used simfactory for submission. I'll take a look at what modifications need to happen for it to work without simfactory. Thanks for catching that! And yes, I'll add more documentation relating to these things. Apologies that I'm behind on that; it's application/grant season, but I've finally come back to life a bit this week.

As a side note: I also typically use simfactory to submit my jobs, but I have a few cases where the simfactory folder is not available. For example, I often I copy the data I need on my computer for faster iteration. When I do so, I don't copy the simfactory folders. Also, when I give my simulation outputs to other people I typically restructure the folders to be simpler and contain only what is needed.
