### 2023-12-04 #6032 @deborahferguson
Ah okay I just tried downloading it from Zenodo and I get an error as well. I'll investigate that further. For what it's worth, it shouldn't explicitly rely on having the simfactory folder as we also don't always keep that, and it's supposed to search for parfiles in the output directories. But clearly something isn't entirely working in some cases so I'll check it out.
