+++ docs/html/_sources/source/mayawaves.utils.rst.txt
@@ -1,3 +1,5 @@
+.. _utils:
+
 mayawaves.utils package
 =======================
 
