+++ docs/html/_sources/source/getting_started.rst.txt
@@ -1,0 +1,182 @@
+.. Mayawaves documentation master file, created by
+   sphinx-quickstart on Thu May  6 10:28:27 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Getting Started
+=====================================
+
+How to cite
+-------------------------------------
+If you use this library for your research please cite it using the following:
+
+bibtex::
+
+    @software{Ferguson_mayawaves_2023,
+        author = {Ferguson, Deborah and Anne, Surendra and Gracia-Linares, Miguel and Iglesias, Hector and Jan, Aasim and Martinez, Erick and Lu, Lu and Meoni, Filippo and Nowicki, Ryan and Trostel, Max and Tsao, Bing-Jyun and Valorz, Finny},
+        doi = {10.5281/zenodo.10035526},
+        month = oct,
+        title = {{mayawaves}},
+        url = {https://github.com/MayaWaves/mayawaves},
+        version = {v2023.10},
+        year = {2023}
+    }
+
+APA::
+
+    Ferguson, D., Anne, S., Gracia-Linares, M., Iglesias, H., Jan, A., Martinez, E., Lu, L., Meoni, F., Nowicki, R., Trostel, M., Tsao, B., & Valorz, F. (2023). mayawaves (Version v2023.10) [Computer software]. https://doi.org/10.5281/zenodo.10035526
+
+These citations can also be downloaded directly from the github repository.
+
+How to install
+-------------------------------------
+
+This package can be installed either from PyPI or from source.
+
+PyPi::
+
+    pip install mayawaves
+
+Source::
+
+    git clone git@github.com:MayaWaves/mayawaves.git
+    cd mayawaves
+    pip install .
+
+To use it in your python code, put::
+
+    import mayawaves
+
+at the top of your file. You can also import specific parts of mayawaves instead if you would rather such as::
+
+    from mayawaves.utils.postprocessingutils import create_h5_from_simulation
+
+Requirements
+-------------------------------------
+Due to its dependencies, this package requires python < 3.11.
+
+It also relies upon the following python packages which are automatically installed from pip when you install mayawaves::
+
+    numpy, scipy, pandas, romspline, wget, matplotlib, mock, numba, scri
+
+This package has been tested with Einstein Toolkit Release Meitner (released on December 6th, 2023).
+It supports output from the following thorns::
+
+    AHFinderDirect
+    PunctureTracker
+    QuasiLocalMeasures
+    Multipole
+    WeylScal4
+
+In addition, it supports the following thorns that are specific to the MAYA catalog simulations::
+
+    IHSpin
+    ShiftTracker
+    Ylm_Decomp
+    RunStats
+
+Basic overview of structure
+-------------------------------------
+.. image:: figures/structure.png
+
+Coalescence
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+The Coalescence object is the heart of the library. It interacts with an h5 file that contains everything about the
+simulation.
+
+CompactObject
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Each Coalescence object has CompactObjects for each of the black holes involved in the simulation.
+Refer to the "Compact Objects" tutorial for more details on how to use CompactObjects.
+Each Coalescence object has a list of all compact objects as well as specific attributes to access the primary,
+secondary, and final compact objects. Primary refers to the larger of the initial black holes.
+
+RadiationBundle
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Each Coalescence object also has a RadiationBundle that handles all the information pertaining to gravitational
+radiation. This is primarily done behind the scenes and you can access all gravitational wave information directly
[…]
