+++ docs/html/_sources/source/getting_help.rst.txt
@@ -1,0 +1,13 @@
+.. Mayawaves documentation master file, created by
+   sphinx-quickstart on Thu May  6 10:28:27 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Getting Help
+=====================================
+
+If you encounter any bugs or unexpected behavior, please file an issue on the github repository at
+https://github.com/MayaWaves/mayawaves/issues.
+
+Feel free to contact the developers directly at dferg [at] illinois.edu if you have any questions regarding this
+library.
