+++ docs/html/_sources/source/mayawaves.compactobject.rst.txt
@@ -1,3 +1,5 @@
+.. _compactobject:
+
 mayawaves.compactobject module
 ==============================
 
