+++ docs/html/_sources/index.rst.txt
@@ -10,29 +10,12 @@ Welcome to Mayawaves's documentation!
 to use this library to analyze both your own simulations and the simulations in the MAYA Catalog
 (https://cgp.ph.utexas.edu/waveform).
 
-The Coalescence object is the heart of the library. It interacts with an h5 file that contains everything about the
-simulation. Each Coalescence object has CompactObjects for each of the black holes involved in the simulation.
-Refer to the compact_objects tutorial for more details on how to use CompactObjects.
-
-Each Coalescence object also has a RadiationBundle that handles all the information pertaining to gravitational
-radiation. This is primarily done behind the scenes and you can access all gravitational wave information directly
-through Coalescence. Refer to the gravitational_waves tutorial for more information on reading gravitational wave data.
-
-There is also a utility to stitch a raw simulation into the h5 format that the Coalescence object reads. This is
-shown in the creating_h5 tutorial. Coalescence objects can also be exported to the format used by LVK analyses, as shown in
-the exporking_lvk tuttorial.
-
-The catalog_utils tutorial walks through using the Catalog object in the catalogutils module to obtain metadata about
-the MAYA catalog and to download simulations from the MAYA catalog.
-
 .. toctree::
-   :maxdepth: 2
-   :caption: Documentation:
+   :maxdepth: 1
 
-   source/mayawaves.coalescence
-   source/mayawaves.compactobject
-   source/mayawaves.radiation
-   source/mayawaves.utils
+   source/getting_started
+   source/getting_help
+   source/contributing
 
 .. toctree::
    :maxdepth: 1
@@ -44,6 +27,16 @@ the MAYA catalog and to download simulations from the MAYA catalog.
    source/notebooks/exporting_lvk
    source/notebooks/reading_raw_files
    source/notebooks/catalog_utils
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Documentation:
+
+   source/mayawaves.coalescence
+   source/mayawaves.compactobject
+   source/mayawaves.radiation
+   source/mayawaves.utils
+
 
 Indices and tables
 ==================
