### 2023-12-21 #6032 @deborahferguson
@Sbozzolo and @cjoana I've added a decent bit of additional documentation for the package, so please have a look when you have a chance (probably after the holidays of course). I'll continue iterating on it, so please let me know if anything is unclear or should be expanded upon.
