### 2023-12-21 #6032 @deborahferguson
@Sbozzolo it seems the reason the zenodo version is failing for me is due to some python code in the GW150914.rpar file that doesn't seem compatible with recent versions of python so I'll add catches to mayawaves to ensure that doesn't cause mayawaves to fail.  Have you been able to run the tutorials on the sample data included in the repo, a simulation downloaded from the MAYA catalog, or an ETK simulation you've performed?
