### 2024-01-10 f7616e6
Update initial spin and mass properties to use the earliest available data
