### 2024-01-10 f718161
fix initial_dimensionless_spin
