### 2024-01-10 3b80e66
Initial parameters (#16)
* Update initial spin and mass properties to use the earliest available data
