### 2024-01-14 #6032 @Sbozzolo
Sorry for the delay, I was gone for the winter break.

I started looking at the new documentation, and it looks greatly improved!

A small comment:
I would move "How to cite" to a seperate section instead of being the very first step in "Getting started". It is important to be easily to find, but people wont't care about it at the beginning

Also, it would be really nice if you fixed the dependency issues that pin Python to <= 3.10. Recent versions of numpy only support >=3.9, leaving a (relatively) tight window of compatibility.

A more important comment is that I still find the capabilities of mayawaves unclear. If I follow the documentation, I read that is "an analysis library for Einstein Toolkit and MAYA simulations", but what can I do with it? The "Getting Started" page offers no help: it tells me how to the code is structured, describes some of the objects, even sketches a basic workflow, without explaning what mayawaves can do. For instance, I am assuming that mayawaves is for simulations of binary black holes only, but I cannot tell for sure. I recommend being very clear on the capabilities and the goals of the code.

I personally find having a single "features" page extremely useful. It helps people potentially interested in the package, provides discoverability, and clarifies what is supposed to be fully implemented or not.

> @Sbozzolo it seems the reason the zenodo version is failing for me is due to some python code in the GW150914.rpar file that doesn't seem compatible with recent versions of python so I'll add catches to mayawaves to ensure that doesn't cause mayawaves to fail.

You can run `GW150914.rpar` with `python3 GW150914.rpar` to produce a valid `par` file. Then, remove the `rpar` from the folder. `mayawaves` still fails in the same way, so I am guessing that the issue is not with the `rpar`.

Pretty much all the simulations I have use thorns that are not supported by mayawaves (mostly the Lean codes, but also my modified version of AH/QLM codes), and getting the Zenodo data is an easy way for me to double check that `mayawaves` works in more a general case besides the data you provide.
