### 2024-01-18 #6032 @deborahferguson
@Sbozzolo Ah, I see cause of that problem. In all simulations I had tested on (including the gallery examples), IO::out_dir was set to be the same as the name of the parameter file and I hadn't realized that was a parameter that could change. In the zenodo version, out_dir is not the same as the parameter file name. Thanks for finding this edge case! I'll make modifications this week to use the value of out_dir when searching for the data and will let you know once that's pushed to the main branch.
