### 2024-02-15 780bb32
fix assumption of file structure in creating the h5 file from the raw simulation
