### 2024-02-18 #6032 @eloisabentivegna
Hi @deborahferguson, do you have updates regarding the Zenodo version?

@Sbozzolo and @cjoana: are there any other points blocking your way down the checklist?
