### 2024-02-19 74855fd
Restructure functions to avoid assuming file structure of simulation data
