+++ tests/test_postprocessingutils.py
@@ -66,6 +66,17 @@ class TestPostprocessingUtils(TestCase):
         if os.path.exists(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp")):
             shutil.rmtree(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp"))
 
+        if os.path.exists(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp_stitched")):
+            shutil.rmtree(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp_stitched"))
+
+        if os.path.exists(os.path.join(TestPostprocessingUtils.CURR_DIR,
+                               "resources/main_test_simulation/stitched/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67.par")):
+            os.remove(os.path.join(TestPostprocessingUtils.CURR_DIR,
+                                   "resources/main_test_simulation/stitched/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67.par"))
+
+        if os.path.exists(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp.h5")):
+            os.remove(os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp.h5"))
+
     def test__simulation_name(self):
         from mayawaves.utils.postprocessingutils import _simulation_name
         expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
@@ -82,21 +93,238 @@ class TestPostprocessingUtils(TestCase):
         actual = _simulation_name("a/b/c/test_name")
         self.assertEqual(expected, actual)
 
-    def test__parameter_file_name_base(self):
-        from mayawaves.utils.postprocessingutils import _parameter_file_name_base
-        # same as simulation name
-        expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
-        actual = _parameter_file_name_base(
-            os.path.join(TestPostprocessingUtils.CURR_DIR,
-                         "resources/main_test_simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"))
-        self.assertEqual(expected, actual)
+    # def test__parameter_file_name_base(self):
+    #     from mayawaves.utils.postprocessingutils import _parameter_file_name_base
+    #     # same as simulation name
+    #     expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
+    #     actual = _parameter_file_name_base(
+    #         os.path.join(TestPostprocessingUtils.CURR_DIR,
+    #                      "resources/main_test_simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"))
+    #     self.assertEqual(expected, actual)
+    #
+    #     # different from simulation name
+    #     expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
+    #     actual = _parameter_file_name_base(
+    #         os.path.join(TestPostprocessingUtils.CURR_DIR,
+    #                      "resources/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67_different_name"))
+    #     self.assertEqual(expected, actual)
 
-        # different from simulation name
-        expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
-        actual = _parameter_file_name_base(
-            os.path.join(TestPostprocessingUtils.CURR_DIR,
-                         "resources/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67_different_name"))
[…]
