### 2024-02-20 3adf7d6
handle different types of quotation marks in parameter file
