### 2024-02-20 2490150
File structure (#17)
* fix assumption of file structure in creating the h5 file from the raw simulation
* Fix issues with parameter file not being generated
