+++ tests/test_postprocessingutils.py
@@ -93,22 +93,6 @@ class TestPostprocessingUtils(TestCase):
         actual = _simulation_name("a/b/c/test_name")
         self.assertEqual(expected, actual)
 
-    # def test__parameter_file_name_base(self):
-    #     from mayawaves.utils.postprocessingutils import _parameter_file_name_base
-    #     # same as simulation name
-    #     expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
-    #     actual = _parameter_file_name_base(
-    #         os.path.join(TestPostprocessingUtils.CURR_DIR,
-    #                      "resources/main_test_simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"))
-    #     self.assertEqual(expected, actual)
-    #
-    #     # different from simulation name
-    #     expected = "D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67"
-    #     actual = _parameter_file_name_base(
-    #         os.path.join(TestPostprocessingUtils.CURR_DIR,
-    #                      "resources/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67_different_name"))
-    #     self.assertEqual(expected, actual)
-
     def test__get_parameter_file_name_and_content(self):
         from mayawaves.utils.postprocessingutils import _get_parameter_file_name_and_content
         temp_raw_directory = os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/temp_raw")
@@ -131,33 +115,19 @@ class TestPostprocessingUtils(TestCase):
         self.assertIsNone(parameter_file_name_base)
         self.assertIsNone(parfile_dict)
 
-        # par file in output-0000
-        shutil.copy2(
-            os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/main_test_simulation/%s/output-0000/%s.par") % (
-                TestPostprocessingUtils.simulation_name,
-                TestPostprocessingUtils.simulation_name), output_0)
-        parameter_file_name_base, parfile_dict = _get_parameter_file_name_and_content(temp_raw_directory)
-        expected_parameter_file = os.path.join(output_0, f'{TestPostprocessingUtils.simulation_name}.par')
-        with open(expected_parameter_file) as f:
-            expected_par_content = f.read()
-        self.assertIsNotNone(parfile_dict)
-        self.assertEqual(TestPostprocessingUtils.simulation_name, parameter_file_name_base)
-        self.assertTrue('par_content' in parfile_dict)
-        self.assertFalse('rpar_content' in parfile_dict)
-        self.assertEqual(expected_par_content, parfile_dict['par_content'])
-
-        # rpar file in output-0000
+        # rpar file in output-0002
         shutil.copy2(
             os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/main_test_simulation/%s/output-0000/%s.rpar") % (
                 TestPostprocessingUtils.simulation_name,
-                TestPostprocessingUtils.simulation_name), output_0)
+                TestPostprocessingUtils.simulation_name), output_2)
         parameter_file_name_base, parfile_dict = _get_parameter_file_name_and_content(temp_raw_directory)
-        expected_rpar_file = os.path.join(output_0, f'{TestPostprocessingUtils.simulation_name}.rpar')
+        expected_rpar_file = os.path.join(output_2, f'{TestPostprocessingUtils.simulation_name}.rpar')
         with open(expected_rpar_file) as f:
             expected_rpar_content = f.read()
-        expected_par_file = os.path.join(TestPostprocessingUtils.CURR_DIR, "resources/main_test_simulation/%s/output-0000/%s.par") % (
-                TestPostprocessingUtils.simulation_name,
-                TestPostprocessingUtils.simulation_name)
[…]
