### 2024-02-21 732762e
lessen size of mp_psi4.h5 data
