### 2024-02-22 #6032 @cjoana
Hi @eloisabentivegna,

Yes, I am waiting to see if there is an update that fixes the issues loading the external data stored in Zenodo.  From a previous comment by @deborahferguson, it seems this will be solved soon.

There are other minor points mentioned before that should be address:
1. For the installation, the issue with the narrow window compatibility of python  versions >=3.9 and <=3.10. If this can not be improved, at least it should be mentioned in the documentation.
2. As suggested by @Sbozzolo, I also think that it should be indicated that the software is dedicated for the analysis of binary black holes, while for other compact objects such neutron stars the methods are not implemented yet (I believe that is the case, but not sure).

After that, from my side, the paper will be ready to go.
