### 2024-02-22 #6032 @deborahferguson
Hi @eloisabentivegna @Sbozzolo @cjoana

I have made changes that should make it compatible with the Zenodo version of GW150914. These changes have been merged into the main branch so you should be able to access them.

I'm currently working on reducing the size of the git repository.

Unfortunately the reason for the <=3.10 compatibility is a dependency of our library not being updated. I'll update the documentation to include a lower limit on the compatibility as well.  To actually improve the compatibility to the most recent python will take time because it will require either removing a little bit of functionality or writing code from scratch to replace the library we're using. I'll also update the documentation to include that it's not yet compatible with neutron stars.

I'll update the pypi distribution once the above changes are finished.
