### 2024-02-23 #6032 @Sbozzolo
> I'll also update the documentation to include that it's not yet compatible with neutron stars.

Please note that my comment on documentation had a larger scope than just mentioning that it doesn't work with neutron stars.

Here:

> A more important comment is that I still find the capabilities of mayawaves unclear. If I follow the documentation, I read that is "an analysis library for Einstein Toolkit and MAYA simulations", but what can I do with it? The "Getting Started" page offers no help: it tells me how to the code is structured, describes some of the objects, even sketches a basic workflow, without explaning what mayawaves can do. For instance, I am assuming that mayawaves is for simulations of binary black holes only, but I cannot tell for sure. I recommend being very clear on the capabilities and the goals of the code.
