### 2024-02-23 #6032 @Sbozzolo
I successfully used mayawaves with the Zenodo data. The tutorials easy to follow and I found that mayawaves produces quantities that tend to roughly agree with kuibit. The code is also very well written, so my key comments are all about documentation: making it clear what mayawaves does, and making it clear how things are computed and what assumptions are being made. I had to dig thorugh the code to understand how physical quantities are calculated, but this should be easy to find in the documentation or the docstrings.

Thanks for your efforts, @deborahferguson!
