### 2024-02-28 #6032 @eloisabentivegna
Thanks everyone for the comments and the thorough reviews.

Regarding the strict Python version requirement, it may be worth exploring other libraries providing the same functionality and compatible with later Python versions. If the library is staying stuck at 3.10, it will hold Mayawaves back and eventually make it unusable with current Python.
