### 2024-03-29 #6032 @cjoana
Hi @eloisabentivegna,

Indeed, I believe we should be close from completion.  I also tested loading the data from Zenodo some time ago, and works well.

@deborahferguson, you mentioned you would update the documentation to clarify the purpose of the library. Is this update still on going?
