### 2024-03-31 #6032 @Sbozzolo
> Hi @Sbozzolo @cjoana, could you complete all items on your checklists, so I know you're done with the review? It feels we're not too far from completion now.

As mentioned in earlier comments, I think there's need for updates in the documentation. For example, to check the box on "Functionalities", the list of available functionalities should be available and be much more granular than "post-processing".

In addition to that, the paper will also need a similar update: the paper does not describe what mayawaves does besides saying that it is a library for "processing, studying, and exporting NR simulations". It does not mention any other code that solves the same problem and how mayawaves differs from them ("Statement of need", and "State of the field"). The paper might also be a little be more specific in terms of compatibility (only a subset of Einstein Toolkit thorns are supported, only binary black holes).
