### 2024-04-01 #6032 @eloisabentivegna
@rhaas80, I'd like to loop you back in at this point. Could you comment on this submission and its suitability for JOSS? I will factor in your viewpoint in the final decision.
