### 2024-04-02 #6032 @Sbozzolo
> @rhaas80, I'd like to loop you back in at this point. Could you comment on this submission and its suitability for JOSS? I will factor in your viewpoint in the final decision.

My take is that the mayawaves is undoubtly suitable for JOSS. The project is not small and the code is well written. People working with the Maya codebase and catalog will certainly benefit from using it, and people working with Einstein Toolkit will have another tool at their disposal.
