+++ tests/resources/main_test_simulation/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/properties.ini
@@ -1,14 +1,0 @@
-
-[properties]
-machine         = frontera-clx
-simulationid    = simulation-D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67-frontera-clx-frontera.tacc.utexas.edu-dferg-2021.01.21-22.28.29-236870
-sourcedir       = /work/05765/dferg/frontera/Cactus
-configuration   = bbh-maya
-configid        = config-bbh-maya-login1.frontera.tacc.utexas.edu-work-05765-dferg-frontera-Cactus
-buildid         = build-bbh-maya-login1.frontera.tacc.utexas.edu-dferg-2020.11.18-22.06.18-145326
-testsuite       = False
-executable      = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/exe/cactus_bbh-maya
-optionlist      = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/cfg/OptionList
-submitscript    = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/run/SubmitScript
-runscript       = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/run/RunScript
-parfile         = /scratch1/05765/dferg/simulations/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67/SIMFACTORY/par/D2.33_q1_a1_0_0_0_a2_0_0_0_m42.67.rpar
