### 2024-04-03 49e5ab9
Reduce filesize (#22)
* Remove example simulation

* lessen size of mp_psi4.h5 data

* reduce file sizes
