### 2024-04-03 1c1adda
Fix error in average run speed in situation where runstats data is unavailable. Closes #19
