### 2024-04-03 f0aee48
Fix error in average run speed in situation where runstats data is unavailable. Closes #19 (#23)
Co-authored-by: Deborah Ferguson <deborahferguson@Debbies-MacBook-Pro.local>
