### 2024-04-03 98bd59c
Add more detailed documentation
