+++ mayawaves/radiation.py
@@ -93,7 +93,7 @@ class RadiationBundle:
     @property
     def extrapolated_sphere(self):
         """The RadiationSphere with extrapolated radius. All :math:`\Psi_4` data has been extrapolated to infinite
-        radius."""
+        radius using the method described in https://arxiv.org/abs/1008.4360 and https://arxiv.org/abs/1108.4421."""
         if self.__extrapolated_sphere is None:
             self.create_extrapolated_sphere()
         return self.__extrapolated_sphere
@@ -136,7 +136,7 @@ class RadiationBundle:
 
     @property
     def included_modes(self) -> list:
-        """:math:`\Psi_4` is decomposed into spherical harmonics labeled by (l, m). This provides a list of all (l,m)
+        """:math:`\Psi_4` is decomposed using spherical harmonics labeled by (l, m). This provides a list of all (l,m)
         modes included."""
         return sorted(list({mode for sphere in self.radiation_spheres.values() for mode in sphere.included_modes}))
 
@@ -274,7 +274,8 @@ class RadiationBundle:
         """Plus component of :math:`rh` for a given mode and extraction radius.
 
         Returns the plus component of strain for a given mode and extraction radius. The strain is the second time
-        integral of :math:`\Psi_4`. If the extraction radius is 0, the data is extrapolated to infinite radius.
+        integral of :math:`\Psi_4` computed using fixed-frequency integration. If the extraction radius is 0, the data
+        is extrapolated to infinite radius.
 
         Args:
             l (int): l value of mode
@@ -300,7 +301,8 @@ class RadiationBundle:
         """Cross component of :math:`rh` for a given mode and extraction radius.
 
         Returns the cross component of strain for a given mode and extraction radius. The strain is the second time
-        integral of :math:`\Psi_4`. If the extraction radius is 0, the data is extrapolated to infinite radius.
+        integral of :math:`\Psi_4` computed using fixed-frequency integration. If the extraction radius is 0, the data
+        is extrapolated to infinite radius.
 
         Args:
             l (int): l value of mode
@@ -353,7 +355,8 @@ class RadiationBundle:
         """Amplitude of :math:`rh` for a given mode and extraction radius.
 
         Returns the amplitude of strain for a given mode and extraction radius. The strain is the second time
-        integral of :math:`\Psi_4`. If the extraction radius is 0, the data is extrapolated to infinite radius.
+        integral of :math:`\Psi_4` computed using fixed-frequency integration. If the extraction radius is 0, the data
+        is extrapolated to infinite radius.
 
         Args:
             l (int): l value of mode
@@ -379,7 +382,8 @@ class RadiationBundle:
         """Phase of :math:`rh` for a given mode and extraction radius.
 
         Returns the phase of strain for a given mode and extraction radius. The strain is the second time
-        integral of :math:`\Psi_4`. If the extraction radius is 0, the data is extrapolated to infinite radius.
+        integral of :math:`\Psi_4` computed using fixed-frequency integration. If the extraction radius is 0, the data
+        is extrapolated to infinite radius.
 
 
         Args:
@@ -430,7 +434,8 @@ class RadiationBundle:
     def get_dEnergy_dt_radiated(self, extraction_radius: float = None, **kwargs) -> tuple:
         """Rate at which energy is radiated, :math:`dE/dt`
 
-        If no lmin, lmax, l, or m are provided, it computes the total sum of all modes.
+        Uses the method described in https://arxiv.org/abs/0707.4654. If no lmin, lmax, l, or m are provided, it
+        computes the total sum of all modes.
 
         Args:
[…]
