### 2024-04-03 2fd69c9
Update jupyter notebooks to tell users where to get the sample data
