+++ docs/html/_sources/source/notebooks/creating_h5.ipynb.txt
@@ -2,6 +2,13 @@
  "cells": [
   {
    "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To obtain the data to run this tutorial, download the tar.gz file from https://zenodo.org/records/10668525 and unzip it. Then set example_simulation to be the path to the resulting directory."
+   ]
+  },
+  {
+   "cell_type": "markdown",
    "metadata": {
     "collapsed": true,
     "jupyter": {
@@ -45,7 +52,7 @@
    },
    "outputs": [],
    "source": [
-    "example_simulation = \"../example_simulation/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33\"\n",
+    "example_simulation = \"D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33\"\n",
     "output_directory = \"\""
    ]
   },
