### 2024-04-03 a969f5f
Add documentation (#24)
* Add more detailed documentation

* Update jupyter notebooks to tell users where to get the sample data
