+++ mayawaves/coalescence.py
@@ -182,8 +182,8 @@ class Coalescence:
     def orbital_angular_momentum_unit_vector(self) -> tuple:
         """Time and orbital angular momentum unit vector over time.
 
-        Unit vector for the orbital angular momentum computed from the positions of the primary and secondary compact
-        objects.
+        Unit vector for the orbital angular momentum computed from the coordinate positions of the primary and secondary
+        compact objects.
         """
         time, separation_vector = self.separation_vector
 
@@ -199,7 +199,7 @@ class Coalescence:
 
     @property
     def orbital_phase_in_xy_plane(self) -> tuple:
-        """Time and orbital phase over time in the xy plane."""
+        """Time and orbital phase over time in the xy plane computed from the coordinate positions."""
         time, separation_vector = self.separation_vector
 
         phase = np.unwrap(np.arctan2(separation_vector[:, 1], separation_vector[:, 0]))
@@ -210,8 +210,8 @@ class Coalescence:
     def orbital_frequency(self) -> tuple:
         """Time and orbital frequency over time.
 
-        Compute and return the orbital frequency as a function of time, calculated from the separation between the
-        primary and secondary compact objects.
+        Compute and return the orbital frequency as a function of time, calculated from the coordinate separation
+        between the primary and secondary compact objects.
         """
         time, separation_vector = self.separation_vector
 
@@ -274,7 +274,7 @@ class Coalescence:
 
     @property
     def merge_time(self) -> float:
-        """Time at which the separation goes below 1e-2"""
+        """Time at which the coordinate separation goes below 1e-2"""
         time, separation_vector = self.separation_vector
         separation_mag = np.linalg.norm(separation_vector, axis=1)
         return time[np.argmax(separation_mag < 1e-2)]
@@ -381,7 +381,7 @@ class Coalescence:
     def recoil_velocity(self, km_per_sec: bool = False) -> np.ndarray:
         """Kick vector of the final object
 
-        Computed from the momentum radiated in gravitational waves
+        Computed from the momentum radiated in gravitational waves.
 
         Args:
             km_per_sec (:obj:`bool`, optional): whether to report the velocity in km/s. Defaults to reporting as a fraction of c.
@@ -405,7 +405,9 @@ class Coalescence:
         return kick_vector
 
     def recoil_speed(self, km_per_sec: bool = False) -> float:
-        """Magnitude of the recoil of the final object
+        """Magnitude of the recoil of the final object.
+
+        Computed from the momentum radiated in gravitational waves.
 
         Args:
             km_per_sec (:obj:`bool`, optional): whether to report the speed in km/s. Defaults to reporting as a fraction of c.
@@ -436,7 +438,7 @@ class Coalescence:
         """Separation at a given time.
 
             Computes and returns the coordinate separation between the primary and secondary compact object centers at
-            the desired time.
+            the desired time. Points from the primary object to the secondary object.
 
             Args:
                  desired_time (float): time at which to return the separation
@@ -458,8 +460,8 @@ class Coalescence:
     def orbital_frequency_at_time(self, desired_time: float) -> float:
         """Orbital frequency at given time.
 
-        Compute and return the orbital frequency at the desired time, calculated from the separation between the
-        primary and secondary compact objects.
+        Compute and return the orbital frequency at the desired time, calculated from the coordinate separation between
+        the primary and secondary compact objects.
 
         Args:
             desired_time (float): time at which to return the orbital separation
@@ -481,8 +483,8 @@ class Coalescence:
     def orbital_angular_momentum_unit_vector_at_time(self, desired_time: float) -> np.ndarray:
         """Orbital angular momentum unit vector at desired time.
 
[…]
