+++ docs/index.rst
@@ -6,8 +6,31 @@
 Welcome to Mayawaves's documentation!
 =====================================
 
-`Mayawaves` serves as an analysis library for Einstein Toolkit and MAYA simulations. Follow the tutorials to learn how
-to use this library to analyze both your own simulations and the simulations in the MAYA Catalog
+`Mayawaves` is a python library for the analysis of numerical relativity simulations of binary black holes performed
+with the Einstein Toolkit (https://einsteintoolkit.org) or provided through the MAYA public waveform catalog
+(https://cgp.ph.utexas.edu/waveform).
+
+Numerical relativity simulations produce large amounts of data and are generally performed over the course of many days.
+The data generated by such simulations is typically broken into multiple output directories, the data within which must
+be stitched together to process the whole simulation. This library is designed to streamline the processing and
+analsis of these simulations.
+
+`Mayawaves` stitches together raw numerical relativity simulations and stores them in h5 files, a format that handles
+numerical data more efficiently than ascii. This reduces the disc space taken by the simulation while still retaining
+the precision of the raw data.
+
+Some of the key functionalities of this library are as follows:
+
+* read and stitch raw numerical relativity data
+* store numerical relativity data in a more efficient format
+* track properties of the black holes (trajectories, spins, masses, horizon information, etc.)
+* compute information about the binary orbit (eccentricity, separation, orbital frequency, kick, etc.)
+* compute quantities pertaining to the gravitational radiation (:math:`\Psi_4`, strain, extrapolated strain, energy radiated, etc.)
+* export numerical relativity data to the format required for LIGO analyses
+* analyze the MAYA waveform catalog
+
+Follow the tutorials to learn how to use this library to analyze both your own simulations and the simulations in the
+MAYA Catalog
 (https://cgp.ph.utexas.edu/waveform).
 
 .. toctree::
