+++ docs/html/_sources/source/notebooks/compact_objects.ipynb.txt
@@ -2,6 +2,15 @@
  "cells": [
   {
    "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "To obtain the data to run this tutorial, download the tar.gz file from https://zenodo.org/records/10668525 and unzip it. Then run the creating_h5 tutorial to create a .h5 file to use with this tutorial. Then set example_h5_filepath to be the path to that .h5 file.\n",
+    "\n",
+    "Once you have run the tutorial on the sample simulation, you can also try downloading a file from the MAYA catalog in the MAYA format (https://cgp.ph.utexas.edu/waveform) and using it with the tutorial."
+   ]
+  },
+  {
+   "cell_type": "markdown",
    "metadata": {
     "collapsed": true,
     "jupyter": {
@@ -47,7 +56,7 @@
    },
    "outputs": [],
    "source": [
-    "example_h5_filepath = \"../example_simulation/D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33.h5\""
+    "example_h5_filepath = \"D11_q5_a1_-0.362_-0.0548_-0.64_a2_-0.0013_0.001_-0.0838_m533.33.h5\""
    ]
   },
   {
