+++ paper.md
@@ -66,9 +66,23 @@ However, these simulations produce vast amounts of data that must be processed i
 Additionally, given the complexity of these simulations, they are typically performed for many days or weeks across many processors, leading to data which is split into several output directories and files.
 Sifting through all this data can be overwhelming for newcomers to the field and is cumbersome for even the most experienced numerical relativists.
 While it is often important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is appropriate.
+There are existing tools to analyze ETK simulations including Kuibit [@Bozzola_kuibit_2021] and SimulationTools [@Hinder], but more versatile tools are still needed.
 
 `Mayawaves` is an open-source python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
 When using the library to interact with a simulation, the user does not need to be familiar with all the types of output files generated by the simulation, but rather, can think in terms of physical concepts such as *coalescences* and *compact objects*.
+Mayawaves stitches together raw numerical relativity simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
+In addition to keeping all the data organized in one place, this also reduces the disc space taken by the simulation while still retaining the precision of the raw data.
+
+Some of the key functionalities of this library are as follows:
+
+* read and stitch raw numerical relativity data
+* store numerical relativity data in a more efficient format
+* track properties of the black holes (trajectories, spins, masses, horizon information, etc.)
+* compute information about the binary orbit (eccentricity, separation, orbital frequency, kick, etc.)
+* compute quantities pertaining to the gravitational radiation ($\Psi_4$, strain, extrapolated strain, energy radiated, etc.)
+* export numerical relativity data to the format required for LIGO analyses
+* analyze the MAYA waveform catalog
+
 
 The `Coalescence` class is the fundamental basis for `Mayawaves`.
 It represents the entirety of the BBH coalescence and serves as the interface between the user and the simulation data.
