### 2024-04-05 #6032 @Sbozzolo
Thank you @deborahferguson! You addressed most of my comments!

The main remaining ones:
- In the paper, for statement of need and state of field: there's other codes that would be worth mentioning (see https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing). I think you should also emphasize what is mayawaves bringing that is not already available and what are the deficits of the other codes compared to mayawaves (e.g., requiring a propritery language or an obsolete version of Python, being unmaintained, ...)
- I think the paper should state that mayawaves is for binary black holes.
- The requested citation for Einstein Toolkit is for the zenodo link (see https://einsteintoolkit.org/citation.html). I would be a nice to also reference papers associated to key codes in Einstein Toolkit (e.g. Cactus, Carpet, ML_BSSN, ...). This can be easily accomplished by adding a few lines describing at the high level Einstein Toolkit (an MAYA).
- Would it be possible to make the dependency on the code that pins Python to an older version optional? Maybe people won't need the features enabled by that package.
