### 2024-04-05 502373d
Update paper text with more detail and more citations
