+++ paper.bib
@@ -525,3 +525,68 @@ year = {2012}
     author = {Ian Hinder and Barry Wardell},
     url = {http://simulationtools.org}
 }
+
+@software{Kastaun,
+    author = {Wolfgang Kastaun},
+    url = {https://github.com/wokast/PyCactus},
+}
+
+@software{Johnson,
+    author = {Daniel Johnson, E. A. Huerta, Roland Haas},
+    url = {https://git.ncsa.illinois.edu/elihu/Gravitational_Waveform_Extractor},
+}
+
+@article{Baumgarte:1998te,
+    author = "Baumgarte, Thomas W. and Shapiro, Stuart L.",
+    title = "{On the numerical integration of Einstein's field equations}",
+    eprint = "gr-qc/9810065",
+    archivePrefix = "arXiv",
+    doi = "10.1103/PhysRevD.59.024007",
+    journal = "Phys. Rev. D",
+    volume = "59",
+    pages = "024007",
+    year = "1998"
+}
+
+@article{PhysRevD.52.5428,
+  title = {Evolution of three-dimensional gravitational waves: Harmonic slicing case},
+  author = {Shibata, Masaru and Nakamura, Takashi},
+  journal = {Phys. Rev. D},
+  volume = {52},
+  issue = {10},
+  pages = {5428--5444},
+  numpages = {0},
+  year = {1995},
+  month = {Nov},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevD.52.5428},
+  url = {https://link.aps.org/doi/10.1103/PhysRevD.52.5428}
+}
+
+@article{Schnetter:2003rb,
+    author = "Schnetter, Erik and Hawley, Scott H. and Hawke, Ian",
+    title = "{Evolutions in 3-D numerical relativity using fixed mesh refinement}",
+    eprint = "gr-qc/0310042",
+    archivePrefix = "arXiv",
+    reportNumber = "AEI-2003-078",
+    doi = "10.1088/0264-9381/21/6/014",
+    journal = "Class. Quant. Grav.",
+    volume = "21",
+    pages = "1465--1488",
+    year = "2004"
+}
+
+@InProceedings{Goodale2002a,
+  author =       {Tom Goodale and Gabrielle Allen and Gerd Lanfermann
+                  and Joan Mass{\'o} and Thomas Radke and Edward
+                  Seidel and John Shalf},
+  title =        {The {Cactus} Framework and Toolkit: Design and
+                  Applications},
+  booktitle =    {Vector and Parallel Processing -- VECPAR'2002, 5th
+                  International Conference, Lecture Notes in Computer
+                  Science},
+  year =         2003,
+  address =      {Berlin},
+  publisher =    {Springer},
+  url =          {http://edoc.mpg.de/3341},
+}
