### 2024-04-05 73d5f91
Update paper text with more detail and more citations
