+++ paper.md
@@ -55,9 +55,11 @@ While analytic solutions exist for the simplest cases within GR, e.g. single BHs
 Approximate methods can be used when the BHs are far apart or have highly unequal masses, but the coalescence of BHs of comparable mass must be solved computationally.
 Numerical relativity (NR) simulations accomplish this by evolving a BBH space-time on supercomputers, enabling us to study the dymanics of BBH systems as well as predict the GWs they emit.
 The Einstein Toolkit (ETK) is a set of tools created to perform these NR simulations [@Loffler:2011ay], and `MAYA` is a branch of ETK used by the MAYA collaboration [@Herrmann_2007; @Jani:2016wkt; @Vaishnav:2007nm; @Healy:2009zm; @Pekowsky:2013ska].
+The Einstein Toolkit is a finite-differencing code, evolved using the BSSN formulation [@Baumgarte:1998te; @PhysRevD.52.5428].
+It is built upon the Cactus infrastructure [@Goodale2002a] with Carpet mesh refinement [@Schnetter:2003rb].
 
 These tools allow us to study the coalescence of compact objects, their evolution, and the gravitational radiation they emit.
-The `Mayawaves` library introduced in this paper is an analysis pipeline used to process and analyze such NR simulations. 
+The `Mayawaves` library introduced in this paper is an analysis pipeline used to process and analyze such NR simulations, specifically for BBHs. 
 
 # Statement of need
 
@@ -66,11 +68,16 @@ However, these simulations produce vast amounts of data that must be processed i
 Additionally, given the complexity of these simulations, they are typically performed for many days or weeks across many processors, leading to data which is split into several output directories and files.
 Sifting through all this data can be overwhelming for newcomers to the field and is cumbersome for even the most experienced numerical relativists.
 While it is often important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is appropriate.
-There are existing tools to analyze ETK simulations including Kuibit [@Bozzola_kuibit_2021] and SimulationTools [@Hinder], but more versatile tools are still needed.
+There are several existing tools to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johsnon], PyCactus [@Kastaun], and SimulationTools [@Hinder].
+A more thorough list of post-processing tools can be found at https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing.
+Each of these tools tackles specific aspects of studying numerical relativity simulations, but more versatile tools are needed.
+Several of these tools require proprietary software and others still require signicant expertise in numerical relativity data.
 
 `Mayawaves` is an open-source python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
 When using the library to interact with a simulation, the user does not need to be familiar with all the types of output files generated by the simulation, but rather, can think in terms of physical concepts such as *coalescences* and *compact objects*.
-Mayawaves stitches together raw numerical relativity simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
+This makes `Mayawaves` unique and particularly powerful, as it dramatically reduces the barrier to entry.
+It is also versatile and easily extensible in order to perform more complex analyses.
+`Mayawaves` stitches together raw numerical relativity simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
 In addition to keeping all the data organized in one place, this also reduces the disc space taken by the simulation while still retaining the precision of the raw data.
 
 Some of the key functionalities of this library are as follows:
