### 2024-04-05 d563f07
Implemented checks on python version before scri related things so that users can use this with the most recent python as long as they don't use those features
