+++ mayawaves/radiation.py
@@ -3,12 +3,10 @@ from enum import Enum
 import h5py
 import numpy as np
 import scipy.integrate
-import scri
 from scipy.ndimage import uniform_filter1d
 from scipy.signal import butter, filtfilt
 from scipy.signal.windows import blackmanharris
 import math
-from spherical_functions import LM_index
 
 
 class Frame(Enum):
@@ -76,6 +74,10 @@ class RadiationBundle:
             center_of_mass (:obj:`numpy.ndarray`, optional): Time series of center of mass. Only necessary if moving to center of mass corrected frame.
 
         """
+        import sys
+        if sys.version_info.major == 3 and sys.version_info.minor > 10:
+            raise ImportError('Unable to change the radiation frame. Python version too recent to be compatible with Scri package. If you would like the ability to move to center-of-mass corrected frame, use python <= 3.10.')
+
         if type(new_frame) != Frame:
             warnings.warn("You must provide the frame as a Frame enum value")
             return
@@ -685,6 +687,11 @@ class RadiationSphere:
             alpha (:obj:`numpy.ndarray`, optional): Offset for center of mass correction. Only necessary if moving to center of mass corrected frame and not providing the center of mass timeseries.
             beta (:obj:`numpy.ndarray`, optional): Boost for center of mass correction. Only necessary if moving to center of mass corrected frame and not providing the center of mass timeseries.
         """
+        import sys
+        if sys.version_info.major == 3 and sys.version_info.minor > 10:
+            raise ImportError(
+                'Unable to change the radiation frame. Python version too recent to be compatible with Scri package. If you would like the ability to move to center-of-mass corrected frame, use python <= 3.10.')
+
         if type(new_frame) != Frame:
             warnings.warn('You must provide the new frame as a Frame enum')
             return
@@ -704,6 +711,11 @@ class RadiationSphere:
         if self.frame == Frame.RAW:
             return self.raw_modes
         if self.frame == Frame.COM_CORRECTED:
+            import sys
+            if sys.version_info.major == 3 and sys.version_info.minor > 10:
+                raise ImportError(
+                    'Unable to return modes in center-of-mass corrected frame. Python version too recent to be compatible with Scri package. If you would like the ability to move to center-of-mass corrected frame, use python <= 3.10.')
+
             if self.__com_corrected_modes is None:
                 self.__frame = Frame.RAW
                 self._generate_com_corrected_modes()
@@ -736,6 +748,11 @@ class RadiationSphere:
     def time(self) -> np.ndarray:
         """Time array associated with all timeseries provided by this RadiationSphere."""
         if self.frame == Frame.COM_CORRECTED:
+            import sys
+            if sys.version_info.major == 3 and sys.version_info.minor > 10:
+                raise ImportError(
+                    'Unable to return time in center-of-mass corrected frame. Python version too recent to be compatible with Scri package. If you would like the ability to move to center-of-mass corrected frame, use python <= 3.10.')
+
             return self.__com_corrected_time
         return self.__time
 
@@ -1239,11 +1256,16 @@ class RadiationSphere:
         extrapolated_sphere = RadiationSphere(mode_dict=temp_modes, time=np.array(self.__time), radius=self.radius,
                                               extrapolated=True)
         if self.frame != Frame.RAW:
+            import sys
+            if sys.version_info.major == 3 and sys.version_info.minor > 10:
+                raise ImportError(
+                    'Unable to set center-of-mass corrected frame. Python version too recent to be compatible with Scri package. If you would like the ability to move to center-of-mass corrected frame, use python <= 3.10.')
+
             extrapolated_sphere.set_frame(self.frame, alpha=self.__alpha, beta=self.__beta)
 
[…]
