### 2024-04-05 e1734c2
update pyproject.toml for current version
