### 2024-04-05 7285418
Python versioning (#25)
* Implemented checks on python version before scri related things so that users can use this with the most recent python as long as they don't use those features

* Update test actions

* update pyproject.toml for current version

* update citation version
