+++ requirements.txt
@@ -5,5 +5,5 @@ romspline
 wget
 matplotlib
 mock
-numba
-scri
+numba; python_version < '3.11'
+scri; python_version < '3.11'
