+++ .github/workflows/test_suite.yml
@@ -4,13 +4,26 @@ on:
     branches: [ main ]
 
 jobs:
-  build:
+  python_3_10_job:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v4
         with:
           python-version: '3.10' 
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      - name: Test with pytest
+        run: python -m unittest discover tests
+  python_3_12_job:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v4
+        with:
+          python-version: '3.12'
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
