### 2024-04-05 #6032 @deborahferguson
Thanks for your suggestions @Sbozzolo

I've updated the paper to include more citations and detail.

I've also modified the code such that users can use it with the most recent python version as long as they don't want to do center-of-mass corrections. In the case they do want to do center-of-mass corrections, it will tell them they need to downgrade to 3.10. Thanks for that suggestion!

I've also uploaded the recent updates to pypi and zenodo as a new version.
