### 2024-04-06 #6032 @Sbozzolo
> Thanks for your suggestions @Sbozzolo
>
> I've updated the paper to include more citations and detail.
>
> I've also modified the code such that users can use it with the most recent python version as long as they don't want to do center-of-mass corrections. In the case they do want to do center-of-mass corrections, it will tell them they need to downgrade to 3.10. Thanks for that suggestion!
>
> I've also uploaded the recent updates to pypi and zenodo as a new version.

Thanks!

I fully ackowledge my bias here, but I think that the paper is not fairly and fully representing the state of the field.

The paper reads
> There are several existing tools to analyze ETK simulations including, but not limited to, Kuibit, Power, PyCactus, and SimulationTools. [...]
Each of these tools tackles specific aspects of studying numerical relativity simulations, but more versatile tools are needed.
Several of these tools require proprietary software and others still require signicant expertise in numerical relativity data.

This implies that existing tools are limited, not versatile enough, and/or not user friendly enough. However, I think that all these problems are already solved by kuibit. Kuibit was also designed around the idea of exposing physical concepts instead of technical details, so I think it's unfair to say that mayawaves is unique in this (as the paragraph below the quoted one does).

While I think I could be perfectly objective in my review, I do reckognize that I have personal interests here in this particular point. So, I'll let @eloisabentivegna comment on this. This is my last point before giving green light.

Thanks again @deborahferguson for all your hard work during this review!
