### 2024-04-08 44dea9c
Update paper to describe what makes mayawaves unique
