+++ paper.md
@@ -68,17 +68,17 @@ However, these simulations produce vast amounts of data that must be processed i
 Additionally, given the complexity of these simulations, they are typically performed for many days or weeks across many processors, leading to data which is split into several output directories and files.
 Sifting through all this data can be overwhelming for newcomers to the field and is cumbersome for even the most experienced numerical relativists.
 While it is often important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is appropriate.
-There are several existing tools to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johsnon], PyCactus [@Kastaun], and SimulationTools [@Hinder].
-A more thorough list of post-processing tools can be found at https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing.
-Each of these tools tackles specific aspects of studying numerical relativity simulations, but more versatile tools are needed.
-Several of these tools require proprietary software and others still require signicant expertise in numerical relativity data.
 
 `Mayawaves` is an open-source python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
+There are other tools to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johsnon], PyCactus [@Kastaun], and SimulationTools [@Hinder], and a more thorough list of post-processing tools can be found at https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing.
+`Mayawaves` builds upon this set of tools, creating a new python library designed for convenience and intuition.
 When using the library to interact with a simulation, the user does not need to be familiar with all the types of output files generated by the simulation, but rather, can think in terms of physical concepts such as *coalescences* and *compact objects*.
-This makes `Mayawaves` unique and particularly powerful, as it dramatically reduces the barrier to entry.
+This dramatically reduces the barrier to entry for the field of NR.
 It is also versatile and easily extensible in order to perform more complex analyses.
-`Mayawaves` stitches together raw numerical relativity simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
+A unique feature of `Mayawaves` is that it stitches together raw NR simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
 In addition to keeping all the data organized in one place, this also reduces the disc space taken by the simulation while still retaining the precision of the raw data.
+`Mayawaves` is also designed with waveform catalogs in mind.
+For this purpose, it interacts effortlessly with the `MAYA` waveform catalog and also has the ability to export ETK simulations to the LVK catalog of NR waveforms.
 
 Some of the key functionalities of this library are as follows:
 
