### 2024-04-08 #6032 @deborahferguson
@Sbozzolo Thanks for your candor, you're correct that it wasn't worded very well. Designing it for intuition and physical concepts are some of my favorite aspects of the library, but you're right that Kuibit is similar in that way and I shouldn't say otherwise. I've done some rephrasing that will hopefully be acceptable. Thanks for all your feedback, it's been very helpful.
