### 2024-04-11 #6032 @eloisabentivegna
Thanks @Sbozzolo for the comments and for disclosing your bias -- I find your viewpoint useful nonetheless. My own perspective is no less biased (I have family connections to SimulationTools), but I have looked at the current wording and find it fair.

It is, however, less descriptive than before. Can we find a description of Mayawaves' unique contributions that we can all agree upon? The creation of Mayawaves must have surely stemmed from shortcomings in the other solutions; the latter should not be understated. @deborahferguson: thanks for the effort so far; could you possibly add a sentence or two explaining your package's unique value to new users?
