+++ paper.md
@@ -69,18 +69,19 @@ NR simulations are crucial for studying BHs and have been instrumental in the de
 However, these simulations produce vast amounts of data that must be processed in order to perform studies, create models, and use them with GW detection pipelines.
 Additionally, given the complexity of these simulations, they are typically performed for many days or weeks across many processors, leading to data which is split into several output directories and files.
 Sifting through all this data can be overwhelming for newcomers to the field and is cumbersome for even the most experienced numerical relativists.
-While it is often important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is appropriate.
+While it can be important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is beneficial.
 
 `Mayawaves` is an open-source python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
-There are other tools to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johsnon], PyCactus [@Kastaun], and SimulationTools [@Hinder], and a more thorough list of post-processing tools can be found at https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing.
-`Mayawaves` builds upon this set of tools, creating a new python library designed for convenience and intuition.
+While other tools exist to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johsnon], PyCactus [@Kastaun], and SimulationTools [@Hinder], `Mayawaves` is unique in the way it not only streamlines simulation analysis but also the production of NR catalogs.
+`Mayawaves` builds upon the existing set of tools, creating a new python library designed for convenience and intuition.
 When using the library to interact with a simulation, the user does not need to be familiar with all the types of output files generated by the simulation, but rather, can think in terms of physical concepts such as *coalescences* and *compact objects*.
 This dramatically reduces the barrier to entry for the field of NR.
 It is also versatile and easily extensible in order to perform more complex analyses.
-A unique feature of `Mayawaves` is that it stitches together raw NR simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
-In addition to keeping all the data organized in one place, this also reduces the disc space taken by the simulation while still retaining the precision of the raw data.
-`Mayawaves` is also designed with waveform catalogs in mind.
-For this purpose, it interacts effortlessly with the `MAYA` waveform catalog and also has the ability to export ETK simulations to the LVK catalog of NR waveforms.
+One of the ways in which `Mayawaves` has uniquely improved the NR analysis infrastructure is that it stitches together raw NR simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
+This significantly reduces the disc space taken by the simulation while still retaining the precision of the raw data.
+It also keeps all the data organized in one place, making it easier to share and distribute simulations.
+`Mayawaves` is designed with waveform catalogs in mind.
+For this purpose, it interacts effortlessly with the `MAYA` waveform catalog and also easily exports ETK simulations to the LVK catalog of NR waveforms.
 
 Some of the key functionalities of this library are as follows:
 
