### 2024-04-16 db3ecff
Add sentence about computational speed
