+++ paper.md
@@ -75,10 +75,11 @@ While it can be important to develop an understanding of these files and their c
 While other tools exist to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], Power [@Johnson], PyCactus [@Kastaun], and SimulationTools [@Hinder], `Mayawaves` is unique in the way it not only streamlines simulation analysis but also the production of NR catalogs.
 `Mayawaves` builds upon the existing set of tools, creating a new python library designed for convenience and intuition, while still being versatile and powerful enough to perform more complex analyses.
 It interacts effortlessly with the `MAYA` waveform catalog and also easily exports ETK simulations to the LVK catalog of NR waveforms.
+The architecture of `Mayawaves` is easily extensible, designed to naturally grow to encompass more types of simulation output.
 One of the ways in which `Mayawaves` has uniquely improved the NR analysis infrastructure is that it stitches together raw NR simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
 This significantly reduces the disc space taken by simulations while still retaining the precision of the raw data.
+By preprocessing the simulations into this format, the library also reduces computational time for future analysis, as the data only has to be stitched once.
 It also keeps all the data organized in one place, making it easier to share and distribute simulations.
-The architecture of `Mayawaves` is easily extensible, designed to naturally grow to encompass more types of simulation output.
 
 When using the library to interact with a simulation, the  user does not need to be familiar with all the types of output files generated by the simulation, but rather, can think in terms of physical concepts such as *coalescences* and *compact objects*.
 This dramatically reduces the barrier to entry for the field of NR.
