+++ paper.bib
@@ -590,3 +590,14 @@ year = {2012}
   publisher =    {Springer},
   url =          {http://edoc.mpg.de/3341},
 }
+
+@article{Ferguson:2023vta,
+    author = "Ferguson, Deborah and others",
+    title = "{Second MAYA Catalog of Binary Black Hole Numerical Relativity Waveforms}",
+    eprint = "2309.00262",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    reportNumber = "UTWI-32-2023",
+    month = "9",
+    year = "2023"
+}
