+++ paper.bib
@@ -52,7 +52,7 @@
 }
 
 @article{LIGOScientific:2014pky,
-    author = "Aasi, J. and others",
+    author = "Aasi, J. et al.",
     collaboration = "LIGO Scientific",
     title = "{Advanced LIGO}",
     eprint = "1411.4547",
@@ -66,7 +66,7 @@
 }
 
 @article{VIRGO:2014yos,
-    author = "Acernese, F. and others",
+    author = "Acernese, F. et al.",
     collaboration = "VIRGO",
     title = "{Advanced Virgo: a second-generation interferometric gravitational wave detector}",
     eprint = "1408.3978",
@@ -81,7 +81,7 @@
 }
 
 @article{eLISA:2013xep,
-    author = "Seoane, Pau Amaro and others",
+    author = "Seoane, Pau Amaro et al.",
     collaboration = "eLISA",
     title = "{The Gravitational Universe}",
     eprint = "1305.5720",
@@ -92,7 +92,7 @@
 }
 
 @article{KAGRA:2020cvd,
-    author = "Akutsu, T. and others",
+    author = "Akutsu, T. et al.",
     collaboration = "KAGRA",
     title = "{Overview of KAGRA : KAGRA science}",
     eprint = "2008.02921",
@@ -104,7 +104,7 @@
 }
 
 @article{LIGOScientific:2016aoc,
-    author = "Abbott, B. P. and others",
+    author = "Abbott, B. P. et al.",
     collaboration = "LIGO Scientific, Virgo",
     title = "{Observation of Gravitational Waves from a Binary Black Hole Merger}",
     eprint = "1602.03837",
@@ -120,7 +120,7 @@
 }
 
 @article{LIGOScientific:2020ibl,
-    author = "Abbott, R. and others",
+    author = "Abbott, R. et al.",
     collaboration = "LIGO Scientific, Virgo",
     title = "{GWTC-2: Compact Binary Coalescences Observed by LIGO and Virgo During the First Half of the Third Observing Run}",
     eprint = "2010.14527",
@@ -135,7 +135,7 @@
 }
 
 @article{LIGOScientific:2018mvr,
-    author = "Abbott, B. P. and others",
+    author = "Abbott, B. P. et al.",
     collaboration = "LIGO Scientific, Virgo",
     title = "{GWTC-1: A Gravitational-Wave Transient Catalog of Compact Binary Mergers Observed by LIGO and Virgo during the First and Second Observing Runs}",
     eprint = "1811.12907",
@@ -151,7 +151,7 @@
 }
 
 @article{LIGOScientific:2021djp,
-    author = "Abbott, R. and others",
+    author = "Abbott, R. et al.",
     collaboration = "LIGO Scientific, VIRGO, KAGRA",
     title = "{GWTC-3: Compact Binary Coalescences Observed by LIGO and Virgo During the Second Part of the Third Observing Run}",
     eprint = "2111.03606",
@@ -163,7 +163,7 @@
 }
 
 @article{LIGOScientific:2017vwq,
-    author = "Abbott, B. P. and others",
+    author = "Abbott, B. P. et al.",
     collaboration = "LIGO Scientific, Virgo",
     title = "{GW170817: Observation of Gravitational Waves from a Binary Neutron Star Inspiral}",
     eprint = "1710.05832",
@@ -179,7 +179,7 @@
 }
 
 @article{LIGOScientific:2020aai,
-    author = "Abbott, B. P. and others",
+    author = "Abbott, B. P. et al.",
     collaboration = "LIGO Scientific, Virgo",
     title = "{GW190425: Observation of a Compact Binary Coalescence with Total Mass $\sim 3.4 M_{\odot}$}",
     eprint = "2001.01761",
@@ -223,7 +223,7 @@
 }
 
[…]
