### 2024-04-16 3fef02f
Fix issues with citations
