+++ paper.bib
@@ -52,7 +52,7 @@
 }
 
 @article{LIGOScientific:2014pky,
-    author = "Aasi, J. et al.",
+    author = "LIGO Scientific Collaboration",
     collaboration = "LIGO Scientific",
     title = "{Advanced LIGO}",
     eprint = "1411.4547",
@@ -65,46 +65,31 @@
     year = "2015"
 }
 
-@article{VIRGO:2014yos,
-    author = "Acernese, F. et al.",
-    collaboration = "VIRGO",
-    title = "{Advanced Virgo: a second-generation interferometric gravitational wave detector}",
-    eprint = "1408.3978",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    doi = "10.1088/0264-9381/32/2/024001",
-    journal = "Class. Quant. Grav.",
-    volume = "32",
-    number = "2",
-    pages = "024001",
-    year = "2015"
+@article{Acernese_2014,
+   title={Advanced Virgo: a second-generation interferometric gravitational wave detector},
+   volume={32},
+   ISSN={1361-6382},
+   url={http://dx.doi.org/10.1088/0264-9381/32/2/024001},
+   DOI={10.1088/0264-9381/32/2/024001},
+   number={2},
+   journal={Classical and Quantum Gravity},
+   publisher={IOP Publishing},
+   author={Acernese, F and Agathos, M and Agatsuma, K and Aisa, D and Allemandou, N and Allocca, A and Amarni, J and Astone, P and Balestri, G and Ballardin, G and Barone, F and Baronick, J-P and Barsuglia, M and Basti, A and Basti, F and Bauer, Th S and Bavigadda, V and Bejger, M and Beker, M G and Belczynski, C and Bersanetti, D and Bertolini, A and Bitossi, M and Bizouard, M A and Bloemen, S and Blom, M and Boer, M and Bogaert, G and Bondi, D and Bondu, F and Bonelli, L and Bonnand, R and Boschi, V and Bosi, L and Bouedo, T and Bradaschia, C and Branchesi, M and Briant, T and Brillet, A and Brisson, V and Bulik, T and Bulten, H J and Buskulic, D and Buy, C and Cagnoli, G and Calloni, E and Campeggi, C and Canuel, B and Carbognani, F and Cavalier, F and Cavalieri, R and Cella, G and Cesarini, E and Mottin, E Chassande- and Chincarini, A and Chiummo, A and Chua, S and Cleva, F and Coccia, E and Cohadon, P-F and Colla, A and Colombini, M and Conte, A and Coulon, J-P and Cuoco, E and Dalmaz, A and D’Antonio, S and Dattilo, V and Davier, M and Day, R and Debreczeni, G and Degallaix, J and Deléglise, S and Pozzo, W Del and Dereli, H and Rosa, R De and Fiore, L Di and Lieto, A Di and Virgilio, A Di and Doets, M and Dolique, V and Drago, M and Ducrot, M and Endrőczi, G and Fafone, V and Farinon, S and Ferrante, I and Ferrini, F and Fidecaro, F and Fiori, I and Flaminio, R and Fournier, J-D and Franco, S and Frasca, S and Frasconi, F and Gammaitoni, L and Garufi, F and Gaspard, M and Gatto, A and Gemme, G and Gendre, B and Genin, E and Gennai, A and Ghosh, S and Giacobone, L and Giazotto, A and Gouaty, R and Granata, M and Greco, G and Groot, P and Guidi, G M and Harms, J and Heidmann, A and Heitmann, H and Hello, P and Hemming, G and Hennes, E and Hofman, D and Jaranowski, P and Jonker, R J G
[…]
