+++ paper.bib
@@ -52,18 +52,18 @@
 }
 
 @article{LIGOScientific:2014pky,
-    author = "LIGO Scientific Collaboration",
-    collaboration = "LIGO Scientific",
-    title = "{Advanced LIGO}",
-    eprint = "1411.4547",
-    archivePrefix = "arXiv",
-    primaryClass = "gr-qc",
-    doi = "10.1088/0264-9381/32/7/074001",
-    journal = "Class. Quant. Grav.",
-    volume = "32",
-    pages = "074001",
-    year = "2015"
-}
+   title={Advanced LIGO},
+   volume={32},
+   ISSN={1361-6382},
+   url={http://dx.doi.org/10.1088/0264-9381/32/7/074001},
+   DOI={10.1088/0264-9381/32/7/074001},
+   number={7},
+   journal={Classical and Quantum Gravity},
+   publisher={IOP Publishing},
+   author={Aasi, J and Abbott, B P and Abbott, R and Abbott, T and Abernathy, M R and Ackley, K and Adams, C and Adams, T and Addesso, P and Adhikari, R X and Adya, V and Affeldt, C and Aggarwal, N and Aguiar, O D and Ain, A and Ajith, P and Alemic, A and Allen, B and Amariutei, D and Anderson, S B and Anderson, W G and Arai, K and Araya, M C and Arceneaux, C and Areeda, J S and Ashton, G and Ast, S and Aston, S M and Aufmuth, P and Aulbert, C and Aylott, B E and Babak, S and Baker, P T and Ballmer, S W and Barayoga, J C and Barbet, M and Barclay, S and Barish, B C and Barker, D and Barr, B and Barsotti, L and Bartlett, J and Barton, M A and Bartos, I and Bassiri, R and Batch, J C and Baune, C and Behnke, B and Bell, A S and Bell, C and Benacquista, M and Bergman, J and Bergmann, G and Berry, C P L and Betzwieser, J and Bhagwat, S and Bhandare, R and Bilenko, I A and Billingsley, G and Birch, J and Biscans, S and Biwer, C and Blackburn, J K and Blackburn, L and Blair, C D and Blair, D and Bock, O and Bodiya, T P and Bojtos, P and Bond, C and Bork, R and Born, M and Bose, Sukanta and Brady, P R and Braginsky, V B and Brau, J E and Bridges, D O and Brinkmann, M and Brooks, A F and Brown, D A and Brown, D D and Brown, N M and Buchman, S and Buikema, A and Buonanno, A and Cadonati, L and Calderón Bustillo, J and Camp, J B and Cannon, K C and Cao, J and Capano, C D and Caride, S and Caudill, S and Cavaglià, M and Cepeda, C and Chakraborty, R and Chalermsongsak, T and Chamberlin, S J and Chao, S and Charlton, P and Chen, Y and Cho, H S and Cho, M and Chow, J H and Christensen, N and Chu, Q and Chung, S and Ciani, G and Clara, F and Clark, J A and Collette, C and Cominsky, L and Constancio, M and Cook, D and Corbitt, T R and Cornish, N and Corsi, A and Costa, C A and Coughlin, M W and Countryman, S and Couvares, P and Coward, D M and Cowart, M J and Coyne, D C and Coyne, R and Craig, K and Creighton, J D E and Creighton, T D and Cripe, J and Crowder, S G and Cumming, A and Cunningham, L and Cutler, C and Dahl, K and Dal Canton, T and Damjanic, M and Danilishin, S L and Danzmann, K and Dartez, L and Dave, I and Daveloza, H and Davies, G S and Daw, E J and DeBra, D and Del Pozzo, W and Denker, T and Dent, T and Dergachev,
[…]
