### 2024-04-16 #6032 @deborahferguson
I've done some additional rephrasing and added some text to make sure it is clear what the unique benefits of Mayawaves are, and I believe the second and third paragraphs of the Statement of Need describe it well. Part of the challenge has been that Mayawaves and Kuibit were being developed at the same time, clearly with a lot of the same ideas, and I want to be sure to give an accurate and fair representation of both.

I appreciate all of your help to improve this paper and library!
