### 2024-04-22 #6032 @eloisabentivegna
Thanks @deborahferguson. I understand the complexities of creating software in a very active field. The new text seems fair to me. @Sbozzolo, could you share your thoughts?

@cjoana, it appears you are done with your checklist -- can you confirm your recommendation?
