### 2024-04-23 #6032 @Sbozzolo
Hi @eloisabentivegna, please read the follow keeping my bias in mind, and feel free to ignore it if you think my position it too biased. :)

I think the new text is fair, but I don't think it answers the question "Do the authors describe how this software compares to other commonly-used packages?": the other packages are just mentioned and nothing is said about them.

If I were to discuss the state of the field in software for post-processing Einstein Toolkit simulations, I would mention:
- SimulationTools: it's rich in features and documentation, but hasn't really seen updates for 6 years and it requires a proprietary language.
- Wolfgang Kastaun's PyCactus: it is also rich in features (including utilities and convenience tools) but it has no documentation and it seems unmaintained.
- POWER, it is officially part of the Einstein Toolkit, it only deals with gravitational waves.
- kuibit, it is officially part of the Einstein Toolkit, designed with user/developer/maintainer friendliness in mind, can work with all sorts of data (from grid variables to timers), thoroughly documented, comes with several examples ready to be used.
- There's also other packages listed in the [wiki](https://docs.einsteintoolkit.org/et-docs/Analysis_and_post-processing) that I am not faimiliar with. For instance, `watpy` seems to be similar to `mayawaves` in working with catalogs (the CoRe one), `rugutils` seems to work with grid variables.

I think the more accurate representation of the state of the field is that there was an underlying issue with usability and user-friendlines that both `kuibit` and `mayawaves` tackled approximately at the same time. The directions ended up being slightly different: kuibit is a general purpose library, but it doesn't implement anything to work with catalogs and it is less sophisticated in the analysis of gravitational waves, areas where mayawaves shines.
