### 2024-04-25 bdb172d
Update version number in CITATION and pyproject.toml
