+++ CITATION.cff
@@ -30,7 +30,7 @@ authors:
   - family-names: "Valorz"
     given-names: "Finny"
 title: "mayawaves"
-version: v2024.4
-date-released: 2024-04-05
+version: v2024.4.25
+date-released: 2024-04-25
 url: "https://github.com/MayaWaves/mayawaves"
-doi: 10.5281/zenodo.10035526
+doi: 10.5281/zenodo.8339252
