### 2024-04-28 #6032 @eloisabentivegna
@Sbozzolo, thanks for the detailed viewpoint. The description you give is extremely useful and will remain on the record for this review.

At the same time, the Statement of Need and, to a lesser extent, the State of the Field often end up being somewhat subjective to who writes them and when. Others may not agree on the state-of-the-art assessment and the feeling that specific functionality is needed, and the review process is not meant to resolve these divergences but to confirm that the authors have fully documented their motivations for writing a given package at a given point in time. This now seems to have been done.

If this is the last point on your list, can you provide your final recommendation for Mayawaves?
