### 2024-04-29 #6032 @Sbozzolo
Yes, @eloisabentivegna, I recommend mayawaves for publication in JOSS and I wholeheartely welcome this new user-friendly public code to the community's shared toolkit.
