+++ paper.bib
@@ -6,7 +6,8 @@
     primaryClass = "gr-qc",
     reportNumber = "LIGO-T1500606",
     month = "3",
-    year = "2017"
+    year = "2017",
+    doi = "10.48550/arXiv.1703.01076"
 }
 
 @article{Healy:2017zqj,
@@ -85,7 +86,8 @@
       year={2020},
       eprint={2008.02921},
       archivePrefix={arXiv},
-      primaryClass={gr-qc}
+      primaryClass={gr-qc},
+      doi={10.1093/ptep/ptaa120}
 }
 
 @article{LIGOScientific:2016aoc,
@@ -493,16 +495,24 @@ year = {2012}
 }
 
 @softward{Hinder,
+    title = {SimulationTools},
     author = {Ian Hinder and Barry Wardell},
     url = {http://simulationtools.org}
 }
 
-@software{Kastaun,
-    author = {Wolfgang Kastaun},
-    url = {https://github.com/wokast/PyCactus},
+@software{2021ascl.soft07017K,
+       author = {{Kastaun}, Wolfgang},
+        title = "{PyCactus: Post-processing tools for Cactus computational toolkit simulation data}",
+ howpublished = {Astrophysics Source Code Library, record ascl:2107.017},
+         year = 2021,
+        month = jul,
+          eid = {ascl:2107.017},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft07017K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @software{Johnson,
+    title = {POWER},
     author = {Daniel Johnson, E. A. Huerta, Roland Haas},
     url = {https://git.ncsa.illinois.edu/elihu/Gravitational_Waveform_Extractor},
 }
@@ -560,6 +570,7 @@ year = {2012}
   address =      {Berlin},
   publisher =    {Springer},
   url =          {http://edoc.mpg.de/3341},
+  doi = {10.1007/3-540-36569-9_13}
 }
 
 @article{ferguson2023second,
@@ -568,7 +579,8 @@ year = {2012}
       year={2023},
       eprint={2309.00262},
       archivePrefix={arXiv},
-      primaryClass={gr-qc}
+      primaryClass={gr-qc},
+      doi = {10.48550/arXiv.2309.00262}
 }
 
 @article{Jan:2023raq,
@@ -578,7 +590,8 @@ year = {2012}
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
     month = "12",
-    year = "2023"
+    year = "2023",
+    doi = "10.48550/arXiv.2312.10241"
 }
 
 @article{tsao2024black,
@@ -587,5 +600,6 @@ year = {2012}
       year={2024},
       eprint={2404.09924},
       archivePrefix={arXiv},
-      primaryClass={gr-qc}
+      primaryClass={gr-qc},
+      doi = {10.48550/arXiv.2404.09924}
 }
