### 2024-05-02 #6032 @deborahferguson
Hi @eloisabentivegna,

Excellent! Thanks for all your work on this. I've added all the missing DOIs I could find. I still can't find ones for SimulationTools, PyCactus, or POWER.
