### 2024-05-02 #6032 @rhaas80
POWER: [https://doi.org/10.1088/1361-6382/aa9cad](https://doi.org/10.1088/1361-6382/aa9cad) (CQG publication) source code (currently) [https://github.com/NCSAGravity/Gravitational_Waveform_Extractor](https://github.com/NCSAGravity/Gravitational_Waveform_Extractor)

SimulationTools: Ian may now what to use

PyCactus: no idea, Wolfgang may have something other than the source code repo. Astrophysics code library has this https://ascl.net/2107.017
