### 2024-05-03 0b2cd0c
Update citation for POWER
