+++ paper.bib
@@ -511,10 +511,18 @@ year = {2012}
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@software{Johnson,
-    title = {POWER},
-    author = {Daniel Johnson, E. A. Huerta, Roland Haas},
-    url = {https://git.ncsa.illinois.edu/elihu/Gravitational_Waveform_Extractor},
+@article{Johnson_2018,
+doi = {10.1088/1361-6382/aa9cad},
+url = {https://dx.doi.org/10.1088/1361-6382/aa9cad},
+year = {2017},
+month = {dec},
+publisher = {IOP Publishing},
+volume = {35},
+number = {2},
+pages = {027002},
+author = {Daniel Johnson and E A Huerta and Roland Haas},
+title = {Python Open source Waveform ExtractoR (POWER): an open source, Python package to monitor and post-process numerical relativity simulations},
+journal = {Classical and Quantum Gravity}
 }
 
 @article{Baumgarte:1998te,
