### 2024-05-03 #6032 @deborahferguson
Thanks @rhaas80! I updated the citation for POWER
