### 2024-05-07 #6032 @eloisabentivegna
Thanks @rhaas80 for filling the gaps, and @deborahferguson for updating the references.

Unless @ianhinder and @wkast object, I would leave the remaining two references as they are, as I can't find a recommended DOId publication to cite for either package.
