### 2024-05-08 #6032 @ianhinder
> Unless @ianhinder and @wkast object, I would leave the remaining two references as they are, as I can't find a recommended DOId publication to cite for either package.

It is fine as it is; there is no other reference for SimulationTools.
