### 2024-05-08 MayaWaves/mayawaves#28 @eloisabentivegna
Small fixes to paper text.
Hello,

I recommend the following changes to the main text:

- Replace "dymanics" -> "dynamics" on line 30
- If an acronym is introduced, use it throughout (e.g. NR, GW, BH)
- "python" -> "Python", "ascii" -> "ASCII", "h5" -> "HDF5" wherever they occur

(Part of JOSS review at https://github.com/openjournals/joss-reviews/issues/6032)
