### 2024-06-10 85d1ae8
Final edits to paper text. Closes #28
