+++ paper.md
@@ -55,7 +55,7 @@ Extracting these signals from noise and using them to infer the parameters of co
 
 While analytic solutions exist for the simplest cases within GR, e.g. single BHs, merging BBHs have no analytic solution.
 Approximate methods can be used when the BHs are far apart or have highly unequal masses, but the coalescence of BHs of comparable mass must be solved computationally.
-Numerical relativity (NR) simulations accomplish this by evolving a BBH space-time on supercomputers, enabling us to study the dymanics of BBH systems as well as predict the GWs they emit.
+Numerical relativity (NR) simulations accomplish this by evolving a BBH space-time on supercomputers, enabling us to study the dynamics of BBH systems as well as predict the GWs they emit.
 The Einstein Toolkit (ETK) is a set of tools created to perform these NR simulations [@L_ffler_2012], and `MAYA` is a branch of ETK used by the MAYA collaboration [@Herrmann_2007; @Jani:2016wkt; @Vaishnav:2007nm; @Healy:2009zm; @Pekowsky:2013ska; @ferguson2023second].
 The Einstein Toolkit is a finite-differencing code, evolved using the BSSN formulation [@Baumgarte:1998te; @PhysRevD.52.5428].
 It is built upon the Cactus infrastructure [@Goodale2002a] with Carpet mesh refinement [@Schnetter:2003rb].
@@ -71,12 +71,12 @@ Additionally, given the complexity of these simulations, they are typically perf
 Sifting through all this data can be overwhelming for newcomers to the field and is cumbersome for even the most experienced numerical relativists.
 While it can be important to develop an understanding of these files and their complexities, in many situations, a simpler, more streamlined workflow is beneficial.
 
-`Mayawaves` is an open-source python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
+`Mayawaves` is an open-source Python library for processing, studying, and exporting NR simulations performed using ETK and `MAYA`.
 While other tools exist to analyze ETK simulations including, but not limited to, Kuibit [@Bozzola_kuibit_2021], POWER [@Johnson_2018], PyCactus [@2021ascl.soft07017K], and SimulationTools [@Hinder], `Mayawaves` is unique in the way it not only streamlines simulation analysis but also the production of NR catalogs.
-`Mayawaves` builds upon the existing set of tools, creating a new python library designed for convenience and intuition, while still being versatile and powerful enough to perform more complex analyses.
+`Mayawaves` builds upon the existing set of tools, creating a new Python library designed for convenience and intuition, while still being versatile and powerful enough to perform more complex analyses.
 It interacts effortlessly with the `MAYA` waveform catalog and also easily exports ETK simulations to the LVK catalog of NR waveforms.
 The architecture of `Mayawaves` is easily extensible, designed to naturally grow to encompass more types of simulation output.
-One of the ways in which `Mayawaves` has uniquely improved the NR analysis infrastructure is that it stitches together raw NR simulations and stores them in h5 files, a format that handles numerical data more efficiently than ascii. 
+One of the ways in which `Mayawaves` has uniquely improved the NR analysis infrastructure is that it stitches together raw NR simulations and stores them in HDF5 files, a format that handles numerical data more efficiently than ASCII. 
 This significantly reduces the disc space taken by simulations while still retaining the precision of the raw data.
 By preprocessing the simulations into this format, the library also reduces computational time for future analysis, as the data only has to be stitched once.
 It also keeps all the data organized in one place, making it easier to share and distribute simulations.
@@ -85,34 +85,34 @@ When using the library to interact with a simulation, the  user does not need to
 This dramatically reduces the barrier to entry for the field of NR.
 Some of the key functionalities of this library are as follows:
 
-* read and stitch raw numerical relativity data
[…]
