+++ paper.md
@@ -18,6 +18,7 @@ authors:
   - name: Hector Iglesias
     affiliation: 1
   - name: Aasim Jan
+    orcid: 0000-0003-2050-7231
     affiliation: 1
   - name: Erick Martinez
     affiliation: 1
