+++ CITATION.cff
@@ -13,6 +13,7 @@ authors:
     given-names: "Hector"
   - family-names: "Jan"
     given-names: "Aasim"
+    orcid: "https://orcid.org/0000-0003-2050-7231"
   - family-names: "Martinez"
     given-names: "Erick"
   - family-names: "Lu"
