+++ CITATION.cff
@@ -31,7 +31,7 @@ authors:
   - family-names: "Valorz"
     given-names: "Finny"
 title: "mayawaves"
-version: v2024.4.25
-date-released: 2024-04-25
+version: v2024.6
+date-released: 2024-06-10
 url: "https://github.com/MayaWaves/mayawaves"
 doi: 10.5281/zenodo.8339252
