+++ CITATION.cff
@@ -11,6 +11,7 @@ authors:
     orcid: "https://orcid.org/0000-0002-5033-2973"
   - family-names: "Iglesias"
     given-names: "Hector"
+    orcid: "https://orcid.org/0009-0008-1089-9239"
   - family-names: "Jan"
     given-names: "Aasim"
     orcid: "https://orcid.org/0000-0003-2050-7231"
