+++ CITATION.cff
@@ -11,8 +11,10 @@ authors:
     orcid: "https://orcid.org/0000-0002-5033-2973"
   - family-names: "Iglesias"
     given-names: "Hector"
+    orcid: "https://orcid.org/0009-0008-1089-9239"
   - family-names: "Jan"
     given-names: "Aasim"
+    orcid: "https://orcid.org/0000-0003-2050-7231"
   - family-names: "Martinez"
     given-names: "Erick"
   - family-names: "Lu"
@@ -30,7 +32,7 @@ authors:
   - family-names: "Valorz"
     given-names: "Finny"
 title: "mayawaves"
-version: v2024.4.25
-date-released: 2024-04-25
+version: v2024.6
+date-released: 2024-06-10
 url: "https://github.com/MayaWaves/mayawaves"
 doi: 10.5281/zenodo.8339252
