+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "mayawaves"
-version = "2024.4.25"
+version = "2024.6"
 authors = [
     { name="Deborah Ferguson", email="deborah.l.ferguson@gmail.com" },
     { name="Surendra Anne" },
