+++ paper.md
@@ -16,6 +16,7 @@ authors:
     orcid: 0000-0002-5033-2973
     affiliation: 1
   - name: Hector Iglesias
+    orcid: 0009-0008-1089-9239
     affiliation: 1
   - name: Aasim Jan
     orcid: 0000-0003-2050-7231
