### 2024-06-10 #6032 @deborahferguson
Hi @eloisabentivegna  I believe I've completed all the tasks listed above.

Release version on GitHub: v2024.6
DOI of release on Zenodo: 10.5281/zenodo.11551465

The authors and license appear to match.
